refactor: restructure routing and state management for improved navigation and category handling
- Updated routing paths to use a centralized configuration in `reactRouter`.
- Introduced a new `CategorySlice` for managing category state in the store.
- Refactored `AppRouter` to utilize dynamic route paths.
- Enhanced `CategoryPage` to fetch categories from the store instead of directly from the API.
- Updated various components to align with new state management and routing structure.
- Removed obsolete `package-lock.json` and unnecessary ESLint configuration.
