+++ .eslintrc.cjs
@@ -1,5 +1,3 @@
-const path = require('path');
-
 module.exports = {
   root: true,
   env: { browser: true, es2020: true },
