+++ src/app/store/category/createCategorySlice.types.ts
@@ -1,0 +1,8 @@
+import type { CategoryModel } from '@/entities/Category';
+
+export interface CategorySlice {
+  categories: CategoryModel[];
+  isLoading: boolean;
+  isError: boolean;
+  requestCategories: () => Promise<void>;
+}
