+++ src/app/store/category/index.ts
@@ -1,0 +1,2 @@
+export { createCategorySlice } from './createCategorySlice';
+export type { CategorySlice } from './createCategorySlice.types';
