+++ src/app/store/rootStore.ts
@@ -3,6 +3,7 @@ import { devtools } from 'zustand/middleware';
 import { immer } from 'zustand/middleware/immer';
 
 import { createAnswersSlice } from './answers';
+import { createCategorySlice } from './category';
 import { createGameSlice } from './game';
 import type { AppStore } from './rootStore.types';
 
@@ -13,6 +14,7 @@ export const useAppStore = create<AppStore>()(
         return {
           game: createGameSlice(...args),
           answers: createAnswersSlice(...args),
+          category: createCategorySlice(...args),
         };
       },
       { name: 'app-storage' }
