+++ src/entities/Question/index.ts
@@ -1,4 +1,10 @@
 export { getQuestions, type RequestQuestionsParams } from './api/getQuestions';
+export { getCorrectAnswers } from './lib/getCorrectAnswers';
 
-export { limitQuestions } from './model/config';
-export type { QuestionModel, Difficulty, Answers } from './model/types';
+export { LIMIT_QUESTIONS } from './model/config';
+export type {
+  QuestionModel,
+  Difficulty,
+  Answers,
+  AnswersByQuestionId,
+} from './model/types';
