+++ src/pages/Category/ui/CategoryList/CategoryList.types.ts
