+++ src/shared/lib/isArraysEqual/index.ts
@@ -1,0 +1,1 @@
+export { isArraysEqual } from './isArraysEqual';
