--- src/shared/lib/react-router/config.ts
@@ -1,6 +1,0 @@
-export const pathKeys = {
-  home: '/',
-  category: 'category/',
-  difficulty: 'difficulty/',
-  result: 'game-result/',
-};
