--- src/shared/lib/react-router/index.ts
@@ -1,1 +1,0 @@
-export { pathKeys } from './config';
