+++ src/shared/lib/reactRouter/config.ts
@@ -1,0 +1,26 @@
+export const routePaths = {
+  home: '/',
+  category: '/category',
+  difficulty: '/category/:categoryName',
+  game: '/category/:categoryName/difficulty/:difficultyLevel',
+  result: '/category/:categoryName/difficulty/:difficultyLevel/game-result',
+  notFound: '*',
+};
+
+export const pathKeys = {
+  home: () => {
+    return routePaths.home;
+  },
+  category: () => {
+    return routePaths.category;
+  },
+  difficulty: (categoryName = '') => {
+    return `/category/${categoryName}`;
+  },
+  game: (categoryName = '', difficultyLevel = '') => {
+    return `/category/${categoryName}/difficulty/${difficultyLevel}`;
+  },
+  result: (categoryName = '', difficultyLevel = '') => {
+    return `/category/${categoryName}/difficulty/${difficultyLevel}/game-result`;
+  },
+};
