+++ src/shared/lib/reactRouter/index.ts
@@ -1,0 +1,1 @@
+export { pathKeys, routePaths } from './config';
