+++ src/shared/ui/couscous/couscous.svg.tsx
@@ -1,0 +1,1866 @@
+import { SVGProps } from 'react';
+import { motion } from 'framer-motion';
+
+export const SmileCouscousSvg = (props: SVGProps<SVGSVGElement>) => {
+  return (
+    <svg
+      viewBox="0 0 350 447"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+      {...props}
+    >
+      <g id="Basic Stance">
+        <g id="Left Arm">
+          <path
+            id="Vector"
+            d="M14.4599 296.27C8.86985 281.27 7.07985 275.27 10.8799 269.27C14.6799 263.27 19.5999 262.56 19.5999 262.56L27.8799 280.67L14.4599 296.27Z"
+            fill="#F14975"
+          />
+          <path
+            id="Vector_2"
+            d="M49.62 246.129C43.62 342.729 69.75 403.129 69.75 403.129C69.75 403.129 79.31 403.129 84.34 405.129C89.37 407.129 98.09 441.359 97.59 446.389H24.47C17.93 446.389 -2.20004 321.609 15.41 274.309C15.41 274.309 6.68996 223.829 17.25 200.689C27.81 177.549 49.62 246.129 49.62 246.129Z"
+            fill="#F14975"
+          />
+          <path
+            id="Vector_3"
+            d="M87.7 405.289C84.01 401.929 68.07 389.519 68.07 389.519C68.07 389.519 43.59 342.729 49.62 246.129C49.62 246.129 27.82 177.529 17.25 200.679C9.39999 217.879 12.25 250.199 14.16 265.679C26.76 273.179 40.93 316.679 47.46 370.219C51.02 399.479 51.65 426.359 49.78 446.409H98.43C98.93 441.349 91.7 408.979 87.7 405.289Z"
+            fill="#D23060"
+          />
+          <path
+            id="Vector_4"
+            d="M81.66 425.25C80.82 422.73 77.8 420.55 69.08 415.85L63.71 407.8L65.8699 417.32L65.67 438.72L67.57 433.91L75.7299 434.02L77.7299 438.38L77.17 433.38L77.2899 429.58L76.2899 432.27L68.91 431.93C68.91 431.93 66.91 429.16 68 420.45C71 420.58 77.59 421.14 79.14 423.6C81.14 426.79 82.83 445.74 82.83 445.74H88.2C88.2 445.74 82.5 427.76 81.66 425.25Z"
+            fill="#9C2950"
+          />
+          <path
+            id="Vector_5"
+            d="M49.0001 289.909C48.5501 276.319 48.6801 261.699 49.6501 246.129C49.6501 246.129 27.8501 177.529 17.2801 200.679C12.1701 211.869 11.5801 229.459 12.2801 244.679C22.5901 266.049 36.0001 282.359 49.0001 289.909Z"
+            fill="#9C2950"
+          />
+          <path
+            id="Vector_6"
+            d="M38.81 442.24C40.5 441.81 42.38 443.62 43.25 446.38H36.53C36.62 444.26 37.47 442.59 38.81 442.24Z"
+            fill="#D23060"
+          />
+          <path
+            id="Vector_7"
+            d="M29.43 442.98C30.34 442.83 31.33 444.26 31.96 446.41H27.96C28.09 444.46 28.62 443.08 29.43 442.98Z"
+            fill="#D23060"
+          />
+        </g>
+        <g id="Right Arm">
+          <path
+            id="Vector_8"
+            d="M341 297.19C345.67 283.3 347.16 277.7 344 272.1C340.84 266.5 336.71 265.88 336.71 265.88L329.8 282.67L341 297.19Z"
+            fill="#F14975"
+          />
+          <path
+            id="Vector_9"
+            d="M318.81 185.6C346 250.51 347 365.22 330.38 426.11C330.38 426.11 330.88 446.4 324.38 446.4H275C267.46 446.4 256.39 435.67 256.89 429.63C257.39 423.59 268.97 390.38 278.02 387.36C287.07 384.34 292.61 388.36 292.61 388.36C292.61 388.36 303.18 325.97 302.68 312.36C302.18 298.75 305.19 247.96 300.68 232.36C296.17 216.76 318.81 185.6 318.81 185.6Z"
+            fill="#D23060"
+          />
+          <path
+            id="Vector_10"
+            d="M318.81 185.6C318.81 185.6 296.17 216.8 300.69 232.39C303.54 242.18 303.41 265.85 303.04 285.48C324.85 251.57 330.59 222.8 328.84 216.02C326.222 205.655 322.87 195.49 318.81 185.6Z"
+            fill="#9C2950"
+          />
+          <path
+            id="Vector_11"
+            d="M330.38 426.11C341.23 387.06 347.24 326.64 338.31 269.59C322.76 279.2 306.42 329.96 299.86 392.32C297.929 410.281 297.001 428.336 297.08 446.4H324.34C330.88 446.4 330.38 426.11 330.38 426.11Z"
+            fill="#F14975"
+          />
+          <path
+            id="Vector_12"
+            d="M272.05 419.05C273.1 416.48 276.4 414.48 286.7 410.05L294.61 400.41L290.43 412.17L289.51 440L287.39 433.29L278.93 432.46L276.47 436.54L277.47 431.44L277.71 427.44L278.49 430.44L286.57 431.64C287.273 426.338 287.544 420.987 287.38 415.64C283.88 415.39 276.53 415.17 274.69 417.64C272.38 420.75 269.27 439.53 269.27 439.53L264.48 438.17C264.48 438.17 271 421.56 272.05 419.05Z"
+            fill="#9C2950"
+          />
+          <path
+            id="Vector_13"
+            d="M309.48 442.26C307.67 441.83 305.65 443.64 304.72 446.4H311.91C311.82 444.28 310.91 442.61 309.48 442.26Z"
+            fill="#D23060"
+          />
+          <path
+            id="Vector_14"
+            d="M319.51 443C318.51 442.85 317.51 444.28 316.81 446.43H321.05C321 444.48 320.38 443.1 319.51 443Z"
+            fill="#D23060"
+          />
+          <path
+            id="Vector_15"
+            d="M327.2 286.38C326.746 290.789 326.519 295.218 326.52 299.65C326.52 323.71 332.99 344.11 341.96 351.25C342.36 344.1 342.58 336.8 342.62 329.41C333.83 324 327.37 306.85 327.2 286.38Z"
+            fill="#D23060"
+          />
+        </g>
+        <g id="Legs">
+          <g id="Left Leg">
+            <path
+              id="Vector_16"
+              d="M112.73 302.91L103 305.59C97.3 340.14 123.47 420 122.8 424C122.13 428 96.63 430 89.92 431.71C83.21 433.42 82.54 446.31 82.54 446.31H153C151.64 441.27 120.11 381.74 112.73 302.91Z"
+              fill="#F14975"
+            />
+            <path
+              id="Vector_17"
+              d="M88.75 432.15C83.14 434.93 82.54 446.31 82.54 446.31H153C152.71 445.23 151 441.68 148.55 436.01C138.67 432.73 125.07 430.71 110.06 430.71C102.932 430.692 95.811 431.173 88.75 432.15Z"
+              fill="#FF99C0"
+            />
+            <path
+              id="Vector_18"
+              d="M122.49 360.11C117.843 341.308 114.581 322.19 112.73 302.91L103 305.59C101 317.73 102.92 335.45 106.24 353.52C111.957 354.887 117.436 357.109 122.49 360.11Z"
+              fill="#D23060"
+            />
+            <path
+              id="Vector_19"
+              d="M105.72 446.31C104.77 440.31 106.42 433.52 107.98 428.84C106.47 429.09 104.98 429.32 103.43 429.55C99.6575 434.342 97.5134 440.214 97.3101 446.31H105.72Z"
+              fill="#F14975"
+            />
+            <path
+              id="Vector_20"
+              d="M129.84 424.56C126.238 427.018 123.224 430.243 121.014 434.003C118.805 437.764 117.454 441.966 117.06 446.31H125.23C124.28 436.3 129.84 424.56 129.84 424.56Z"
+              fill="#F14975"
+            />
+          </g>
+          <g id="Right Leg">
+            <path
+              id="Vector_21"
+              d="M230.25 296.87C237.96 332.76 232.15 390.12 225.1 409.58C218.05 429.04 211.69 442 212 446.31H281.44C281.44 446.31 277.86 438.42 270.44 434.4C263.02 430.38 248.19 431.27 245.17 426.57C242.15 421.87 251.54 381.06 246.06 303.36C245.29 291.6 230.25 296.87 230.25 296.87Z"
+              fill="#F14975"
+            />
+            <path
+              id="Vector_22"
+              d="M233.1 361C237.83 355.35 242.82 351.11 247.87 348.57C247.79 335.23 247.31 320.17 246.12 303.4C245.29 291.64 230.25 296.91 230.25 296.91C234.2 315.26 234.6 339.29 233.1 361Z"
+              fill="#D23060"
+            />
+            <path
+              id="Vector_23"
+              d="M214.17 437.86C212.72 441.86 211.89 444.67 212.02 446.31H281.46C281.46 446.31 277.88 438.42 270.46 434.4C265.08 431.47 255.74 431.14 249.89 429.4H249.55C234.82 429.37 221.81 432.74 214.17 437.86Z"
+              fill="#FF99C0"
+            />
+            <path
+              id="Vector_24"
+              d="M248 446.31C249 440.31 247.29 433.52 245.7 428.84C247.24 429.09 248.8 429.32 250.34 429.55C254.182 434.313 256.375 440.194 256.59 446.31H248Z"
+              fill="#F14975"
+            />
+            <path
+              id="Vector_25"
+              d="M223.43 424.56C223.43 424.56 235.25 431.62 236.43 446.31H228.09C229.1 436.3 223.43 424.56 223.43 424.56Z"
+              fill="#F14975"
+            />
+          </g>
+        </g>
+        <motion.g
+          id="Ears &#38; Horns"
+          animate={{ y: [0, 50, 0] }}
+          transition={{
+            duration: 2,
+            repeat: Infinity,
+            ease: 'easeInOut',
+            type: 'keyframes',
+            times: [0, 1],
+          }}
+        >
+          <g id="Group 25">
+            <path
+              id="Vector_26"
+              d="M81.0725 0C58.6425 7.31 23.0725 63.4 74.2425 90.22C89.7325 89.47 100.243 79.13 103.023 61.94C62.0525 40.96 81.0725 0 81.0725 0Z"
+              fill="#FBD8A8"
+            />
+            <path
+              id="Vector_27"
+              d="M82.6226 74.2198C66.1026 74.2198 52.6226 55.6098 51.9626 32.2998C44.1926 51.7898 46.0526 75.4498 74.2626 90.2198C89.2626 89.4998 99.5926 79.7398 102.763 63.4498C97.3526 70.1098 90.3126 74.2198 82.6226 74.2198Z"
+              fill="#D39F7A"
+            />
+            <path
+              id="Vector_28"
+              opacity={0.22}
+              d="M84.6227 85.9599C67.5027 85.9599 53.0627 68.2599 48.6227 44.1099C45.9227 60.6999 51.5627 78.3298 74.2527 90.2198C79.6317 90.0536 84.8632 88.4216 89.3826 85.4999C87.8142 85.8039 86.2203 85.9579 84.6227 85.9599Z"
+              fill="#FBD8A8"
+            />
+            <path
+              id="Vector_29"
+              opacity={0.6}
+              d="M73.0426 25.2501C73.0426 15.3801 76.0426 6.48012 80.7226 0.120117C72.6426 2.92012 62.9426 11.9201 56.2026 23.5401C58.8026 41.2501 66.4826 54.0901 75.5526 54.0901C77.8564 54.0466 80.0869 53.2727 81.9226 51.8801C76.4926 45.4601 73.0426 35.9101 73.0426 25.2501Z"
+              fill="#FFF2E0"
+            />
+            <g
+              id="Group"
+              style={{
+                mixBlendMode: 'multiply',
+              }}
+              opacity={0.33}
+            >
+              <path
+                id="Vector_30"
+                d="M89.7124 52.68C88.5512 51.5792 87.4528 50.4141 86.4224 49.1899C81.0624 51.9199 78.6324 57.85 76.8424 64.49C81.5324 56.44 82.6224 53.87 89.7124 52.68Z"
+                fill="#D39F7A"
+              />
+              <path
+                id="Vector_31"
+                d="M56.6224 76.3399C57.6399 77.604 58.7248 78.8125 59.8724 79.9599C62.7324 54.4999 64.0524 53.9598 73.0024 44.5898C64.0924 49.7298 57.8024 60.3999 56.6224 76.3399Z"
+                fill="#D39F7A"
+              />
+              <path
+                id="Vector_32"
+                d="M77.8425 32.1601C77.3696 30.1858 77.0355 28.1809 76.8425 26.1601C72.4325 25.9001 67.7125 29.5501 62.9425 33.8701C70.1325 28.9401 73.3025 28.6901 77.8425 32.1601Z"
+                fill="#D39F7A"
+              />
+              <path
+                id="Vector_33"
+                d="M67.0025 17.98C61.203 22.6143 55.9612 27.9065 51.3824 33.75C49.3969 38.9896 48.2593 44.5122 48.0125 50.11C54.0325 30.95 57.1625 27.37 67.0025 17.98Z"
+                fill="#D39F7A"
+              />
+            </g>
+          </g>
+          <g id="Group 26">
+            <path
+              id="Vector_34"
+              d="M267.972 0C290.402 7.31 325.972 63.4 274.802 90.22C259.312 89.47 248.802 79.13 246.022 61.94C286.992 40.96 267.972 0 267.972 0Z"
+              fill="#FBD8A8"
+            />
+            <path
+              id="Vector_35"
+              d="M266.422 74.2198C282.942 74.2198 296.422 55.6098 297.082 32.2998C304.852 51.7898 302.992 75.4498 274.782 90.2198C259.782 89.4998 249.452 79.7398 246.282 63.4498C251.692 70.1098 258.732 74.2198 266.422 74.2198Z"
+              fill="#D39F7A"
+            />
+            <path
+              id="Vector_36"
+              opacity={0.22}
+              d="M264.422 85.9599C281.542 85.9599 295.982 68.2599 300.422 44.1099C303.122 60.6999 297.482 78.3298 274.792 90.2198C269.413 90.0536 264.182 88.4216 259.662 85.4999C261.231 85.8039 262.825 85.9579 264.422 85.9599Z"
+              fill="#FBD8A8"
+            />
+            <path
+              id="Vector_37"
+              opacity={0.6}
+              d="M276.002 25.2501C276.002 15.3801 273.052 6.48012 268.322 0.120117C276.402 2.92012 286.102 11.9201 292.842 23.5401C290.242 41.2501 282.562 54.0901 273.492 54.0901C271.189 54.0466 268.958 53.2727 267.122 51.8801C272.552 45.4601 276.002 35.9101 276.002 25.2501Z"
+              fill="#FFF2E0"
+            />
+            <g
+              id="Group_2"
+              style={{
+                mixBlendMode: 'multiply',
+              }}
+              opacity={0.33}
+            >
+              <path
+                id="Vector_38"
+                d="M259.332 52.68C260.491 51.5767 261.589 50.4116 262.622 49.1899C267.982 51.9199 270.412 57.85 272.202 64.49C267.512 56.44 266.422 53.87 259.332 52.68Z"
+                fill="#D39F7A"
+              />
+              <path
+                id="Vector_39"
+                d="M292.422 76.3399C291.405 77.604 290.32 78.8125 289.172 79.9599C286.312 54.4999 284.992 53.9598 276.042 44.5898C284.952 49.7298 291.242 60.3999 292.422 76.3399Z"
+                fill="#D39F7A"
+              />
+              <path
+                id="Vector_40"
+                d="M271.202 32.1601C271.675 30.1858 272.009 28.1809 272.202 26.1601C276.612 25.9001 281.332 29.5501 286.102 33.8701C278.912 28.9401 275.742 28.6901 271.202 32.1601Z"
+                fill="#D39F7A"
+              />
+              <path
+                id="Vector_41"
+                d="M282.042 17.98C287.842 22.6143 293.084 27.9065 297.662 33.75C299.648 38.9896 300.785 44.5122 301.032 50.11C295.012 30.95 291.882 27.37 282.042 17.98Z"
+                fill="#D39F7A"
+              />
+            </g>
+          </g>
+        </motion.g>
+        <motion.g
+          id="Torso"
+          clipPath="url(#clip0_69_18457)"
+          animate={{ y: [0, 50, 0] }}
+          transition={{
+            duration: 2,
+            repeat: Infinity,
+            ease: 'easeOut',
+            type: 'keyframes',
+            times: [0, 1],
+          }}
+        >
+          <g id="Torso_2">
+            <path
+              id="Vector_42"
+              d="M34.3581 231.242C42.5715 249.985 47.9237 256.113 47.9237 256.113C47.9237 256.113 51.8553 251.107 51.8553 244.939C51.8553 244.939 54.7064 296.474 82.9279 314.857L79.7066 292.509C79.7066 292.509 85.779 309.811 91.8515 314.136L92.5718 304.764C92.5718 304.764 97.2137 320.264 99.7147 322.787C102.216 325.31 102.926 310.892 102.926 310.892C102.926 310.892 110.069 325.671 112.57 327.113C115.071 328.555 114.711 322.427 114.711 322.427C114.711 322.427 120.423 329.275 122.214 330.357C124.005 331.438 123.995 327.834 123.995 327.834C123.995 327.834 130.067 338.287 132.928 340.089C135.789 341.891 136.5 324.59 136.5 324.59C136.5 324.59 139.711 333.24 141.852 332.88C143.993 332.52 146.494 323.148 146.494 323.148C146.494 323.148 149.705 328.154 151.136 328.915C152.566 329.676 153.997 323.148 153.997 323.148C153.997 323.148 156.138 331.438 157.568 332.52C158.999 333.601 165.782 318.101 165.782 318.101C165.782 318.101 166.492 332.52 166.142 334.682C165.792 336.845 182.208 322.427 182.208 322.427C182.208 322.427 180.067 331.438 182.208 332.159C184.349 332.88 189.711 324.229 189.711 324.229C189.711 324.229 187.56 340.45 190.061 340.089C192.562 339.729 210.07 323.508 213.631 315.939C213.631 315.939 213.281 334.322 215.782 334.682C218.283 335.043 230.418 319.904 230.418 314.857C230.418 314.857 227.567 324.59 229.707 325.31C231.848 326.031 243.283 315.218 244.714 310.171C244.714 310.171 243.643 319.543 246.504 319.543C249.365 319.543 253.647 309.09 253.647 309.09C253.647 309.09 258.289 314.857 259.36 314.497C260.43 314.136 261.861 302.962 261.861 302.962C261.861 302.962 267.213 316.299 265.792 320.985C264.372 325.671 284 300.079 286.861 286.031C286.861 286.031 287.931 295.753 287.571 297.195C287.211 298.637 294.004 289.625 297.215 276.659C297.215 276.659 297.925 283.508 299.006 283.508C300.086 283.508 306.499 266.206 305.788 257.194C309.36 266.206 306.859 270.892 307.929 270.211C309 269.53 312.931 261.92 312.931 254.711L316.503 257.955C316.503 257.955 320.074 248.223 318.644 241.014C318.644 241.014 320.074 248.944 321.145 249.304C322.215 249.665 328.998 230.921 322.215 220.468C322.215 220.468 326.497 225.474 327.927 223.712C329.358 221.95 327.567 204.247 323.286 198.841C323.286 198.841 327.927 202.846 329.718 204.608C331.509 206.37 327.217 194.515 326.147 192.352C325.076 190.19 334 204.247 334.71 203.166C335.42 202.085 335.07 184.062 328.998 177.944L335.07 176.142C335.07 176.142 330.789 161.003 325.787 157.398L330.068 158.119C330.068 158.119 331.859 142.98 323.996 135.41L327.567 137.213C327.567 137.213 326.147 117.388 314.002 110.9C301.857 104.411 315.072 110.9 315.072 110.9L320.784 111.26C320.784 111.26 310.07 102.609 301.857 102.609L305.788 97.2022C305.788 97.2022 292.213 91.0745 283.639 90.3535C275.066 89.6326 279.718 83.8654 279.718 83.8654C277.019 82.7044 274.047 82.3296 271.144 82.784C266.142 83.5049 263.641 68.3758 264.362 66.5735C265.082 64.7712 258.649 68.3758 258.649 68.3758C258.649 68.3758 256.859 60.4458 257.929 59.3644C258.999 58.283 247.575 63.6899 247.575 63.6899C247.575 63.6899 237.571 58.283 235.43 54.6785C233.289 51.0739 235.79 47.1089 235.79 47.1089C235.79 47.1089 227.927 48.9112 228.647 53.2367C229.367 57.5621 219.003 51.7948 220.784 46.388C222.565 40.9812 214.721 48.1903 215.782 52.8762C216.842 57.5621 205.067 52.8762 203.637 47.4694L198.995 48.9112C198.995 48.9112 180.417 39.5393 174.705 40.2603C174.705 40.2603 176.846 48.1903 170.784 46.388C164.721 44.5857 160.419 41.7021 155.777 43.1439C155.777 43.1439 156.138 47.149 151.136 46.388C147.94 45.9537 144.71 45.833 141.492 46.0275C141.492 46.0275 137.21 53.5971 129.707 51.7948C122.204 49.9926 122.214 46.388 122.214 46.388C122.214 46.388 116.852 51.3943 119.713 56.8412C122.574 62.2881 112.92 62.248 111.499 57.9226C110.079 53.5971 106.848 56.4808 105.777 62.248C104.707 68.0153 93.2821 64.7712 93.2821 64.7712C93.2821 64.7712 90.4209 66.934 95.423 69.0967C100.425 71.2594 94.7127 72.7012 88.2801 74.103C81.8474 75.5048 80.777 79.1093 80.777 79.1093C85.0687 78.028 79.7066 84.8766 71.1431 83.4348C62.5796 81.993 59.3583 87.4399 59.3583 87.4399C59.3583 87.4399 64.7105 85.9981 64.3603 88.1608C64.0102 90.3235 47.5735 89.6026 40.0805 99.3349C42.8243 99.9867 45.7052 99.7338 48.2938 98.614C48.2938 98.614 45.0825 102.939 45.0825 104.742C45.0825 104.742 31.507 116.997 30.7967 124.206L35.4386 125.648C35.4386 125.648 25.7947 134.66 25.0744 137.543C24.3541 140.427 35.0784 137.904 35.0784 137.904C35.0784 137.904 26.8651 141.909 24.7242 147.996L28.6458 148.717C28.6458 148.717 18.2816 160.953 17.2112 169.243L23.9939 167.801C23.9939 167.801 16.5009 179.696 15.0703 187.616L22.2132 182.61C22.2132 182.61 18.6418 199.191 19.3521 203.517C20.0623 207.842 26.8551 196.668 26.8551 196.668C26.8551 196.668 24.5041 223.041 29.166 238.721C30.2365 242.255 34.3581 231.242 34.3581 231.242Z"
+              fill="url(#paint0_radial_69_18457)"
+            />
+            <path
+              id="Vector_43"
+              d="M335.06 176.092C335.06 176.092 334.85 175.331 334.46 174.089C331.485 174.042 328.534 174.632 325.806 175.821C331.329 179.376 333.76 189.038 334.81 196.397C334.36 190.31 332.809 181.719 328.978 177.864L335.06 176.092Z"
+              fill="#31134F"
+            />
+            <path
+              id="Vector_44"
+              d="M330.208 155.836C322.825 150.32 313.952 152.122 313.952 152.122C317.543 152.012 325.076 158.83 331.069 165.198C329.548 161.884 327.717 158.74 325.776 157.348L330.058 158.069C330.058 158.069 330.128 157.218 330.208 155.836Z"
+              fill="#31134F"
+            />
+            <path
+              id="Vector_45"
+              d="M313.952 133.758C319.744 134.489 323.956 137.854 330.048 150.249C329.578 145.393 328.117 139.335 323.986 135.36L327.557 137.163C327.557 137.163 327.477 135.951 327.157 134.039C322.906 132.496 318.265 132.397 313.952 133.758Z"
+              fill="#31134F"
+            />
+            <path
+              id="Vector_46"
+              d="M279.708 83.8156C279.305 83.63 278.89 83.4695 278.467 83.335C274.296 85.5077 273.705 90.5841 281.469 91.8958C284.86 92.388 288.304 92.415 291.703 91.9759C289.057 91.1639 286.34 90.604 283.589 90.3038C275.106 89.5829 279.708 83.8156 279.708 83.8156Z"
+              fill="#31134F"
+            />
+            <path
+              id="Vector_47"
+              d="M311.011 105.082C308.173 103.59 305.048 102.726 301.847 102.549L305.779 97.1423L304.438 96.5615C300.076 99.0146 298.516 104.421 298.516 104.421C302.167 103.53 306.709 104.081 311.011 105.082Z"
+              fill="#31134F"
+            />
+            <path
+              id="Vector_48"
+              d="M232.259 322.928C233.168 315.585 233.366 308.171 232.849 300.79C229.178 312.728 223.739 324.049 216.712 334.372C219.834 332.62 227.367 323.278 229.718 317.611C229.047 320.545 228.327 324.81 229.718 325.281C230.015 325.341 230.324 325.302 230.598 325.17L230.808 325.08L230.938 325.02L231.048 324.97C231.393 324.807 231.727 324.624 232.049 324.419C236.14 321.806 243.583 314.167 244.714 310.151C244.714 310.151 243.823 318.001 245.864 319.333L246.865 298.096C242.757 306.799 237.868 315.11 232.259 322.928Z"
+              fill="#31134F"
+            />
+            <path
+              id="Vector_49"
+              d="M320.784 117.097C319.037 114.524 316.711 112.397 313.992 110.889C301.847 104.401 315.062 110.889 315.062 110.889L320.774 111.25C318.607 109.557 316.32 108.025 313.932 106.664C310.465 106.63 307.059 107.578 304.108 109.397C308.75 109.988 315.232 113.513 320.784 117.097Z"
+              fill="#31134F"
+            />
+            <path
+              id="Vector_50"
+              d="M318.203 183.691L329.908 203.066C328.893 199.396 327.63 195.8 326.127 192.302C325.056 190.14 333.98 204.197 334.69 203.116C334.715 203.083 334.732 203.046 334.74 203.006C330.128 191.101 318.203 183.691 318.203 183.691Z"
+              fill="#31134F"
+            />
+            <path
+              id="Vector_51"
+              d="M300.906 279.903C303.127 274.356 306.289 263.643 305.778 257.184C309.35 266.196 306.849 270.882 307.919 270.201C308.182 269.974 308.409 269.707 308.59 269.41C309.42 254.902 301.867 239.933 301.867 239.933C305.008 252.599 302.867 269.4 300.906 279.903Z"
+              fill="#31134F"
+            />
+            <path
+              id="Vector_52"
+              d="M311.441 263.092C312.37 260.376 312.869 257.532 312.921 254.661L316.493 257.905C316.493 257.905 316.543 257.775 316.613 257.545C314.882 249.935 309.47 241.234 309.47 241.234C310.831 244.639 312.121 254.271 311.441 263.092Z"
+              fill="#31134F"
+            />
+            <path
+              id="Vector_53"
+              d="M318.213 251.567C319.015 248.09 319.157 244.494 318.634 240.964C318.634 240.964 320.064 248.894 321.135 249.254C321.335 249.324 321.745 248.704 322.225 247.602C321.472 243.182 319.54 239.048 316.633 235.637C318.394 239.042 318.524 245.69 318.213 251.567Z"
+              fill="#31134F"
+            />
+            <path
+              id="Vector_54"
+              d="M292.953 288.574C294.753 284.746 296.177 280.752 297.205 276.649C297.205 276.649 297.915 283.497 298.996 283.497C299.146 283.497 299.416 283.117 299.756 282.496C299.446 272.163 293.813 259.617 293.813 259.617C295.616 269.219 295.323 279.097 292.953 288.574Z"
+              fill="#31134F"
+            />
+            <path
+              id="Vector_55"
+              d="M275.826 308.83C280.598 301.521 285.52 292.529 286.831 286.021C286.831 286.021 287.901 295.743 287.541 297.185C287.441 297.586 287.881 297.185 288.631 296.184C289.632 284.539 285.96 271.713 285.96 271.713C286.11 283.968 281.208 297.586 275.826 308.83Z"
+              fill="#31134F"
+            />
+            <path
+              id="Vector_56"
+              d="M316.183 192.643C320.955 195.827 325.316 207.561 328.298 217.554C327.867 211.386 326.127 202.385 323.296 198.79C323.296 198.79 327.937 202.795 329.728 204.558C330.068 204.908 330.188 204.738 330.148 204.217C324.636 198.29 316.183 192.643 316.183 192.643Z"
+              fill="#31134F"
+            />
+            <path
+              id="Vector_57"
+              d="M309.25 206.31C317.253 213.209 322.475 225.554 325.156 233.504C325.317 228.919 324.626 224.143 322.205 220.418C322.205 220.418 326.487 225.424 327.918 223.662C327.964 223.599 328.004 223.532 328.038 223.462C321.125 212.928 309.25 206.31 309.25 206.31Z"
+              fill="#31134F"
+            />
+            <path
+              id="Vector_58"
+              d="M246.834 319.483C249.676 318.742 253.637 309.08 253.637 309.08C253.637 309.08 257.239 313.555 258.829 314.376L253.547 302.582L246.834 319.483Z"
+              fill="#31134F"
+            />
+            <path
+              id="Vector_59"
+              d="M259.39 314.467C260.45 313.876 261.851 302.952 261.851 302.952C261.851 302.952 267.203 316.289 265.782 320.975C265.572 321.656 265.782 321.696 266.382 321.225C267.793 306.146 260.03 283.548 260.03 283.548L259.39 314.467Z"
+              fill="#31134F"
+            />
+            <path
+              id="Vector_60"
+              d="M35.8787 221.029C35.8787 221.029 30.8767 230.591 29.5361 239.312C30.9767 240.313 34.3581 231.232 34.3581 231.232C37.2893 237.92 39.8503 242.987 41.9412 246.761C38.8399 237.179 35.5086 225.605 35.8787 221.029Z"
+              fill="#31134F"
+            />
+            <path
+              id="Vector_61"
+              d="M82.898 314.697L79.7067 292.499C79.7067 292.499 84.5087 306.156 89.8308 312.234C85.5691 301.891 77.5258 281.775 77.6559 277.45C77.6559 277.45 75.1249 298.326 82.898 314.697Z"
+              fill="#31134F"
+            />
+            <path
+              id="Vector_62"
+              d="M91.361 313.746C91.517 313.882 91.6807 314.009 91.8512 314.126L92.5715 304.755C92.5715 304.755 96.263 317.11 98.844 321.546L92.2514 293.741C92.2514 293.741 90.4506 305.646 91.361 313.746Z"
+              fill="#31134F"
+            />
+            <path
+              id="Vector_63"
+              d="M54.8366 234.796C54.8366 234.796 49.4044 247.282 47.4937 255.572C47.7738 255.933 47.9238 256.103 47.9238 256.103C47.9238 256.103 51.8554 251.097 51.8554 244.929C52.6511 255.162 54.5188 265.284 57.4277 275.127C55.7389 261.749 54.8736 248.28 54.8366 234.796Z"
+              fill="#31134F"
+            />
+            <path
+              id="Vector_64"
+              d="M160.509 328.945C162.91 324.66 165.781 318.091 165.781 318.091C165.781 318.091 166.002 322.487 166.132 326.792L169.433 308.019L160.509 328.945Z"
+              fill="#31134F"
+            />
+            <path
+              id="Vector_65"
+              d="M141.932 332.85C144.053 332.289 146.494 323.137 146.494 323.137C146.494 323.137 149.495 327.853 150.986 328.815C149.195 324.279 145.984 315.798 145.984 313.726C145.944 313.726 141.922 324.159 141.932 332.85Z"
+              fill="#31134F"
+            />
+            <path
+              id="Vector_66"
+              d="M182.999 329.485C185.08 323.528 187.38 316.239 187.38 316.239L166.172 334.372C166.172 334.492 166.172 334.602 166.172 334.692C165.812 336.855 182.238 322.437 182.238 322.437C182.238 322.437 180.098 331.448 182.238 332.169C184.379 332.89 189.741 324.239 189.741 324.239C189.741 324.239 188.961 330.126 188.891 334.722C191.082 328.164 194.103 319.953 196.304 317.27C199.235 313.645 188.581 321.986 182.999 329.485Z"
+              fill="#31134F"
+            />
+            <path
+              id="Vector_67"
+              d="M152.596 327.563C153.189 326.131 153.657 324.65 153.997 323.137C153.997 323.137 155.737 329.886 157.128 331.999L154.127 313.055L152.596 327.563Z"
+              fill="#31134F"
+            />
+            <path
+              id="Vector_68"
+              d="M215.082 309.19L189.411 339.629C189.581 339.949 189.801 340.12 190.071 340.079C192.572 339.719 210.08 323.498 213.641 315.929C213.641 315.929 213.291 334.312 215.792 334.673C215.973 334.687 216.155 334.656 216.322 334.582C217.312 323.949 215.082 309.19 215.082 309.19Z"
+              fill="#31134F"
+            />
+            <path
+              id="Vector_69"
+              d="M113.76 327.153C114.931 326.302 114.711 322.417 114.711 322.417C114.711 322.417 120.423 329.265 122.214 330.347C122.32 330.425 122.442 330.481 122.571 330.512C122.699 330.543 122.833 330.548 122.964 330.527C119.633 322.327 115.111 310.412 114.901 305.666C114.901 305.666 111.479 316.629 113.76 327.153Z"
+              fill="#31134F"
+            />
+            <path
+              id="Vector_70"
+              d="M123.194 330.437C123.995 329.926 123.995 327.824 123.995 327.824C123.995 327.824 129.167 336.725 132.208 339.518C128.206 329.506 122.554 314.697 122.454 309.48C121.313 316.446 121.565 323.569 123.194 330.437Z"
+              fill="#31134F"
+            />
+            <path
+              id="Vector_71"
+              d="M100.335 323.068C102.336 322.857 102.926 310.882 102.926 310.882C102.926 310.882 110.069 325.661 112.57 327.103C112.672 327.169 112.783 327.22 112.9 327.253C109.499 320.154 104.407 308.91 102.986 302.472C101.534 309.252 100.647 316.141 100.335 323.068Z"
+              fill="#31134F"
+            />
+            <path
+              id="Vector_72"
+              d="M26.7749 197.609C26.6248 199.611 26.3147 204.367 26.2646 210.305C26.7548 204.758 27.9153 199.501 30.1662 195.336C28.9968 196.034 27.8648 196.792 26.7749 197.609Z"
+              fill="#31134F"
+            />
+            <path
+              id="Vector_73"
+              d="M134.528 338.838C136.059 335.023 136.529 324.58 136.529 324.58C136.529 324.58 138.77 330.587 140.761 332.4C139.07 327.393 136.569 319.573 136.569 316.58C135.2 323.919 134.517 331.371 134.528 338.838Z"
+              fill="#31134F"
+            />
+            <path
+              id="Vector_74"
+              d="M19.7324 204.328C20.8829 205.329 24.0042 200.993 25.7349 198.43C23.5031 200.146 21.4876 202.126 19.7324 204.328Z"
+              fill="#31134F"
+            />
+            <path
+              id="Vector_75"
+              d="M155.337 48.4005C158.659 47.199 158.809 44.3954 156.488 42.9136C156.248 42.9607 156.011 43.0208 155.777 43.0938C155.777 43.0938 156.137 47.0989 151.135 46.3379C149.55 46.1138 147.954 45.9734 146.354 45.9174C150.525 48.5707 152.166 49.5419 155.337 48.4005Z"
+              fill="#31134F"
+            />
+            <path
+              id="Vector_76"
+              d="M117.582 64.1803C128.827 66.0327 128.487 63.3393 124.125 56.9612C121.574 53.2265 120.703 49.9524 121.124 47.5093C119.733 49.3116 117.632 52.9562 119.663 56.791C122.514 62.1978 112.87 62.1978 111.45 57.8724C110.979 56.4706 110.339 55.8698 109.629 55.7897C109.829 59.164 111.37 63.159 117.582 64.1803Z"
+              fill="#31134F"
+            />
+            <path
+              id="Vector_77"
+              d="M212.721 56.9615C221.344 58.5836 218.723 49.4921 219.844 44.686C218.093 45.7574 215.092 49.8125 215.782 52.8263C216.782 57.1818 206.638 53.4271 204.077 48.5209C203.627 51.8751 204.457 55.3996 212.721 56.9615Z"
+              fill="#31134F"
+            />
+            <path
+              id="Vector_78"
+              d="M106.507 70.7289C106.507 70.7289 96.7735 68.0355 92.9219 65.0317C92.3517 65.7326 91.6414 67.3647 95.423 69.0368C98.9544 70.5587 97.1436 71.7302 93.5122 72.7915C94.7627 72.6307 96.0229 72.5571 97.2837 72.5713C106.678 72.7515 106.507 70.7289 106.507 70.7289Z"
+              fill="#31134F"
+            />
+            <path
+              id="Vector_79"
+              d="M257.689 70.899C258.199 71.0292 261.34 67.8451 264.362 66.4734C264.952 64.7813 258.639 68.3257 258.639 68.3257C258.184 66.2814 257.85 64.2118 257.639 62.1279C255.898 65.6524 254.468 70.098 257.689 70.899Z"
+              fill="#31134F"
+            />
+            <path
+              id="Vector_80"
+              d="M272.765 82.5835C272.235 82.5835 271.695 82.6536 271.134 82.7337C269.934 82.9139 268.874 82.163 267.973 80.9014C268.113 81.5848 268.377 82.2366 268.753 82.8238C270.374 85.1467 271.785 84.0553 272.765 82.5835Z"
+              fill="#31134F"
+            />
+            <path
+              id="Vector_81"
+              d="M224.635 55.7901C231.178 57.6324 229.998 52.9265 232.519 50.2431C233.401 49.3154 234.373 48.477 235.42 47.7399C235.526 47.5059 235.646 47.2786 235.78 47.0591C235.78 47.0591 227.917 48.8614 228.637 53.1868C229.187 56.501 223.215 53.8777 221.214 50.0328C221.214 52.6361 221.994 55.0191 224.635 55.7901Z"
+              fill="#31134F"
+            />
+            <path
+              id="Vector_82"
+              d="M78.8262 84.1656C90.2409 81.1618 94.6026 76.9465 87.0496 78.158C85.0351 78.4732 83.0043 78.6738 80.9671 78.7588C80.8963 78.8921 80.8329 79.0291 80.777 79.1693C85.0688 78.0879 79.7066 84.9366 71.1431 83.4948C70.0393 83.3108 68.9203 83.2337 67.8018 83.2645C69.5325 84.6362 72.8038 85.7576 78.8262 84.1656Z"
+              fill="#31134F"
+            />
+            <path
+              id="Vector_83"
+              d="M248.795 65.5222L257.969 59.1742C257.469 58.9439 247.565 63.6399 247.565 63.6399C247.565 63.6399 246.994 63.3295 246.104 62.8088C246.825 63.8728 247.737 64.793 248.795 65.5222Z"
+              fill="#31134F"
+            />
+            <path
+              id="Vector_84"
+              d="M172.784 49.2419C179.287 49.5222 178.166 43.1341 174.785 40.2305C174.785 40.2305 176.926 48.1605 170.863 46.3582C167.642 45.3569 164.921 44.1454 162.39 43.3544C165.061 46.178 168.873 49.0716 172.784 49.2419Z"
+              fill="#31134F"
+            />
+            <path
+              id="Vector_85"
+              d="M60.3691 86.208C59.9875 86.5695 59.6515 86.9764 59.3687 87.4196C59.3687 87.4196 64.7208 85.9778 64.3707 88.1405C64.0806 89.9227 52.9461 89.7525 44.8228 95.0292C52.2358 93.2269 62.2698 90.9641 68.2723 90.3833C76.1555 89.6123 68.3823 85.9477 60.3691 86.208Z"
+              fill="#31134F"
+            />
+            <path
+              id="Vector_86"
+              d="M24.7344 147.977L28.636 148.698C28.636 148.698 28.1458 149.288 27.3554 150.29C30.6868 148.157 33.8781 146.525 35.8789 146.465C32.0961 146.019 28.2622 146.54 24.7344 147.977Z"
+              fill="#31134F"
+            />
+            <path
+              id="Vector_87"
+              d="M30.1667 165.779C25.7892 165.636 21.433 166.449 17.4015 168.162C17.3215 168.532 17.2614 168.893 17.2114 169.253L23.9942 167.811C23.9942 167.811 21.8233 171.256 19.5924 175.681C23.2239 171.066 27.3656 166.56 30.1667 165.779Z"
+              fill="#31134F"
+            />
+            <path
+              id="Vector_88"
+              d="M29.1659 175.691C29.1659 175.691 19.9722 180.057 15.2803 187.466L22.2131 182.57C22.2131 182.57 20.2123 191.772 19.512 198.25C21.913 189.569 25.4344 179.046 29.1659 175.691Z"
+              fill="#31134F"
+            />
+            <path
+              id="Vector_89"
+              d="M25.0942 138.064C25.8445 140.066 35.0983 137.883 35.0983 137.883C35.0983 137.883 29.0959 140.787 26.0046 145.473C30.4564 141.918 36.0087 138.154 40.6005 137.403C40.5705 137.403 31.857 135.941 25.0942 138.064Z"
+              fill="#31134F"
+            />
+            <path
+              id="Vector_90"
+              d="M43.7319 105.983C42.6614 106.985 41.0508 108.607 39.3301 110.489C43.7419 107.686 48.6138 105.212 53.1857 104.221L43.7319 105.983Z"
+              fill="#31134F"
+            />
+            <path
+              id="Vector_91"
+              d="M30.8167 123.986C30.8111 124.053 30.8111 124.12 30.8167 124.186L35.4586 125.628C35.4586 125.628 33.7679 127.21 31.7571 129.253C35.298 126.651 39.418 124.95 43.7619 124.296C39.4896 123.516 35.1214 123.411 30.8167 123.986Z"
+              fill="#31134F"
+            />
+            <path
+              id="Vector_92"
+              d="M40.451 98.8643C40.3209 99.0144 40.1908 99.1546 40.0708 99.3148C42.8147 99.9666 45.6955 99.7138 48.2841 98.5939C48.2841 98.5939 47.6639 99.435 46.9636 100.526C51.1653 98.8342 55.1169 97.5226 57.5179 97.2622C51.7842 97.0739 46.0497 97.6121 40.451 98.8643Z"
+              fill="#31134F"
+            />
+            <path
+              id="Vector_93"
+              d="M225.976 259.487C222.837 268.263 218.727 276.659 213.721 284.519L212.38 270.742L214.561 276.449L225.976 259.487Z"
+              fill="#31134F"
+            />
+            <path
+              id="Vector_94"
+              d="M269.594 217.834C269.764 235.297 261.881 250.245 261.881 250.245L258.879 236.648L262.401 241.014C266.322 235.186 268.393 227.056 269.594 217.834Z"
+              fill="#31134F"
+            />
+            <path
+              id="Vector_95"
+              d="M149.735 263.793C157.418 276.248 161.74 277.971 161.74 277.971C161.74 277.971 172.744 267.958 175.175 257.845C175.175 257.845 166.892 271.312 162.68 273.185C157.998 271.533 153.716 268.288 149.735 263.793Z"
+              fill="#31134F"
+            />
+            <path
+              id="Vector_96"
+              d="M119.543 272.353C118.866 276.574 118.568 280.846 118.652 285.12C118.652 285.12 108.128 277.51 106.567 267.427C106.567 267.427 110.819 276.389 115.291 278.631L119.543 272.353Z"
+              fill="#31134F"
+            />
+            <path
+              id="Vector_97"
+              d="M79.4966 231.832C86.2093 243.917 88.4502 245.71 88.4502 245.71L91.8016 229.139L87.3298 240.563L79.4966 231.832Z"
+              fill="#31134F"
+            />
+            <path
+              id="Vector_98"
+              d="M59.4185 116.526C73.2241 111.12 81.6675 112.581 81.6675 112.581L64.4905 116.987L76.8655 121.393C69.4425 120.582 61.8995 119.831 59.4185 116.526Z"
+              fill="#31134F"
+            />
+            <path
+              id="Vector_99"
+              d="M57.7978 143.841C51.7953 147.846 44.7925 158.39 44.7925 158.39L60.899 153.914L49.8946 155.035L57.7978 143.841Z"
+              fill="#31134F"
+            />
+            <path
+              id="Vector_100"
+              d="M187.33 96.5913C192.472 73.3018 187.33 64.5508 187.33 64.5508C187.33 64.5508 192.022 74.183 192.252 82.4634C192.293 87.6006 190.553 92.5931 187.33 96.5913Z"
+              fill="#31134F"
+            />
+            <path
+              id="Vector_101"
+              d="M192.582 102.859C179.387 89.2021 181.457 70.869 181.578 71.5198C185.429 86.2884 191.692 101.558 192.582 102.859Z"
+              fill="#31134F"
+            />
+            <path
+              id="Vector_102"
+              d="M249.295 108.016C265.622 109.578 268.083 106.664 268.083 106.664C268.083 106.664 253.997 109.358 252.207 105.322C250.416 101.287 258.019 92.7866 258.019 92.7866C253.147 95.5201 249.936 100.236 249.295 108.016Z"
+              fill="#31134F"
+            />
+            <path
+              id="Vector_103"
+              d="M281.959 139.355C294.704 137.353 305.668 144.512 305.668 144.512C305.668 144.512 290.232 141.378 286.881 141.819C283.529 142.259 288.441 149.879 298.065 153.013C283.749 151.671 281.959 139.355 281.959 139.355Z"
+              fill="#31134F"
+            />
+            <path
+              id="Vector_104"
+              d="M214.731 225.565C210.646 231.657 206.044 237.385 200.976 242.686C200.976 242.686 195.604 232.954 195.273 225.224C195.273 225.224 198.625 235.978 201.976 236.649C205.328 237.319 214.731 225.565 214.731 225.565Z"
+              fill="#31134F"
+            />
+          </g>
+        </motion.g>
+        <motion.g
+          id="Eyes"
+          animate={{ y: [0, 50, 0] }}
+          transition={{
+            duration: 2,
+            repeat: Infinity,
+            ease: 'easeOut',
+            type: 'keyframes',
+            times: [0, 1],
+          }}
+        >
+          <g id="Frame 4">
+            <g id="Group 21">
+              <path
+                id="Vector_105"
+                d="M11.6753 160.702C33.9553 156.042 43.4454 148.152 44.6754 146.182C45.9053 144.212 15.5353 152.762 10.6753 153.462C5.81535 154.162 7.64535 161.542 11.6753 160.702Z"
+                fill="#431A68"
+              />
+              <path
+                id="Vector_106"
+                d="M6.9452 155.371C34.5052 154.761 44.5452 142.771 45.6352 139.371C46.7252 135.971 9.7252 141.151 3.6352 140.661C-2.4548 140.171 1.9652 155.481 6.9452 155.371Z"
+                fill="#8C47CD"
+              />
+              <path
+                id="Vector_107"
+                d="M54.3949 194.092C54.6349 178.922 49.3949 167.512 37.6049 164.892C23.6749 161.792 13.9149 178.402 17.6049 198.552C20.6549 215.332 31.8449 225.492 41.7749 223.822C50.5149 222.362 54.1949 207.272 54.3949 194.092Z"
+                fill="#431A68"
+              />
+              <path
+                id="Vector_108"
+                d="M51.8752 190.342C52.0352 176.692 47.5152 166.432 37.4052 164.062C25.4552 161.252 17.1452 176.162 20.3352 194.292C22.9852 209.382 32.6052 218.532 41.1052 217.052C48.6052 215.752 51.7252 202.222 51.8752 190.342Z"
+                fill="#7E32A8"
+              />
+              <path
+                id="Vector_109"
+                d="M48.795 187.041C49.175 176.371 45.595 168.211 37.195 166.041C27.275 163.511 20.055 174.951 22.385 189.221C24.335 201.111 32.195 208.551 39.315 207.631C45.605 206.841 48.455 196.321 48.795 187.041Z"
+                fill="url(#paint1_radial_69_18457)"
+              />
+              <g
+                id="Vector_110"
+                style={{
+                  mixBlendMode: 'overlay',
+                }}
+              >
+                <path
+                  d="M42.7851 174.341C41.3751 170.691 38.2951 168.201 35.8951 168.761C33.4951 169.321 32.7051 172.761 34.1051 176.391C35.5051 180.021 38.6051 182.531 40.9951 181.961C43.3851 181.391 44.1951 177.981 42.7851 174.341Z"
+                  fill="#D5BBA2"
+                />
+              </g>
+              <g
+                id="Vector_111"
+                style={{
+                  mixBlendMode: 'overlay',
+                }}
+              >
+                <path
+                  d="M45.2352 190.322C45.3324 189.827 45.3272 189.317 45.2198 188.825C45.1125 188.332 44.9053 187.866 44.611 187.457C44.3168 187.047 43.9418 186.703 43.509 186.444C43.0763 186.185 42.5951 186.017 42.0952 185.952C41.0457 185.938 40.0299 186.322 39.2525 187.027C38.475 187.732 37.9936 188.706 37.9052 189.752C37.8079 190.246 37.8132 190.756 37.9205 191.248C38.0279 191.741 38.2351 192.207 38.5293 192.616C38.8235 193.026 39.1986 193.371 39.6313 193.629C40.064 193.888 40.5452 194.056 41.0452 194.122C42.0933 194.131 43.1065 193.745 43.883 193.041C44.6594 192.336 45.1422 191.366 45.2352 190.322Z"
+                  fill="#D5BBA2"
+                />
+              </g>
+            </g>
+          </g>
+          <g id="Frame 5">
+            <g id="Group 22">
+              <path
+                id="Vector_112"
+                d="M333.29 195.89C333.29 180.09 326.05 166.79 314.22 166.79C302.39 166.79 294.54 180.58 294.54 196.37C294.54 212.16 302.4 224 314.22 224C326.04 224 333.29 211.68 333.29 195.89Z"
+                fill="#431A68"
+              />
+              <path
+                id="Vector_113"
+                d="M331.08 192.49C331.08 178.22 324.67 166.22 314.19 166.22C303.71 166.22 296.74 178.66 296.74 192.93C296.74 207.2 303.74 217.93 314.19 217.93C324.64 217.93 331.08 206.75 331.08 192.49Z"
+                fill="#7E32A8"
+              />
+              <path
+                id="Vector_114"
+                d="M313.82 209.59C321.121 209.59 327.04 200.269 327.04 188.77C327.04 177.272 321.121 167.95 313.82 167.95C306.518 167.95 300.6 177.272 300.6 188.77C300.6 200.269 306.518 209.59 313.82 209.59Z"
+                fill="url(#paint2_radial_69_18457)"
+              />
+              <g
+                id="Vector_115"
+                style={{
+                  mixBlendMode: 'overlay',
+                }}
+              >
+                <path
+                  d="M322.32 178.79C320.91 175.15 317.83 172.65 315.43 173.22C313.03 173.79 312.24 177.22 313.65 180.84C315.06 184.46 318.14 186.98 320.53 186.41C322.92 185.84 323.73 182.43 322.32 178.79Z"
+                  fill="#D5BBA2"
+                />
+              </g>
+              <path
+                id="Vector_116"
+                d="M332.42 158.4C314.14 154.58 306.36 148.11 305.31 146.49C304.26 144.87 329.22 151.89 333.21 152.49C337.2 153.09 335.72 159.09 332.42 158.4Z"
+                fill="#431A68"
+              />
+              <path
+                id="Vector_117"
+                d="M336.3 154.04C313.69 153.53 305.45 143.69 304.56 140.88C303.67 138.07 334.02 142.33 339.03 141.93C344.04 141.53 340.39 154.13 336.3 154.04Z"
+                fill="#8C47CD"
+              />
+              <g
+                id="Vector_118"
+                style={{
+                  mixBlendMode: 'overlay',
+                }}
+              >
+                <path
+                  d="M324.77 194.78C324.867 194.285 324.862 193.776 324.755 193.283C324.647 192.79 324.44 192.325 324.146 191.915C323.851 191.506 323.476 191.161 323.044 190.902C322.611 190.643 322.13 190.476 321.63 190.41C320.582 190.396 319.568 190.779 318.79 191.482C318.013 192.185 317.531 193.156 317.44 194.2C317.343 194.695 317.348 195.204 317.455 195.697C317.563 196.19 317.77 196.655 318.064 197.065C318.358 197.474 318.733 197.819 319.166 198.078C319.599 198.337 320.08 198.504 320.58 198.57C321.627 198.582 322.64 198.198 323.417 197.495C324.194 196.793 324.677 195.823 324.77 194.78Z"
+                  fill="#D5BBA2"
+                />
+              </g>
+            </g>
+          </g>
+        </motion.g>
+        <motion.g
+          id="Mouth"
+          clipPath="url(#clip1_69_18457)"
+          animate={{ y: [0, 50, 0] }}
+          transition={{
+            duration: 2,
+            repeat: Infinity,
+            ease: 'easeOut',
+            type: 'keyframes',
+            times: [0, 1],
+          }}
+        >
+          <g id="Group 1">
+            <path
+              id="Vector_119"
+              d="M183.5 167.3C167.32 144.3 146.82 150.67 148.18 166.84C165.49 189.55 188.7 193.27 218.59 175.5C226.57 144.5 201.5 150.89 183.5 167.3Z"
+              fill="#A763CC"
+            />
+            <path
+              id="Vector_120"
+              d="M148.26 163.28C150.42 172.94 166.1 175.92 180.01 162.86C170.63 152.19 160.47 150.4 154.2 153.99C153.013 154.937 151.85 155.957 150.71 157.05C149.363 158.877 148.519 161.025 148.26 163.28Z"
+              fill="#CC80E0"
+            />
+            <path
+              id="Vector_121"
+              d="M188.74 162.96C202.91 171.65 218.86 174.58 219.94 162.43C218.16 148.84 202.58 152.55 188.74 162.96Z"
+              fill="#CC80E0"
+            />
+            <path
+              id="Vector_122"
+              d="M177.58 177.33C166.75 174.92 161.86 165.93 163.68 163.88C165.5 161.83 173.48 168.67 175.76 171.63C178.04 174.59 179.63 177.78 177.58 177.33Z"
+              fill="#491B2A"
+            />
+            <path
+              id="Vector_123"
+              d="M176.92 176.17C174.77 174.17 168.81 170.52 165.6 170.1C168.66 173.806 172.893 176.357 177.6 177.33C177.781 177.375 177.969 177.375 178.15 177.33C177.75 176.95 177.35 176.56 176.92 176.17Z"
+              fill="#9D375B"
+            />
+            <path
+              id="Vector_124"
+              d="M196.69 178.01C194.59 176.61 203.48 160.58 207.45 162.51C211.42 164.44 211.11 169.35 208.18 172.25C205.25 175.15 198.59 179.27 196.69 178.01Z"
+              fill="#491B2A"
+            />
+            <path
+              id="Vector_125"
+              d="M209.66 170.19C207.4 168.31 200.19 173.06 196.39 176.86C196.34 177.43 196.39 177.86 196.69 178.01C198.59 179.27 205.25 175.15 208.18 172.25C208.783 171.65 209.283 170.954 209.66 170.19Z"
+              fill="#9D375B"
+            />
+            <path
+              id="Vector_126"
+              d="M190.86 161.43C188.28 163.223 185.821 165.184 183.5 167.3C179.36 161.4 174.93 157.45 170.65 155.08C183.65 168.95 185.34 176.24 185.34 176.24C186.836 171.182 188.681 166.234 190.86 161.43Z"
+              fill="#72369C"
+            />
+            <path
+              id="Vector_127"
+              d="M204.81 182.37C209.581 180.455 214.189 178.157 218.59 175.5C220.14 169.5 220.44 164.91 219.8 161.5C215.35 171.59 207.8 179.5 204.81 182.37Z"
+              fill="#72369C"
+            />
+            <path
+              id="Vector_128"
+              d="M165.17 178.23C159.299 171.747 154.672 164.239 151.52 156.08C150.28 157.475 149.342 159.11 148.764 160.884C148.185 162.658 147.98 164.533 148.16 166.39C154.87 173.85 165.17 178.23 165.17 178.23Z"
+              fill="#72369C"
+            />
+            <path
+              id="Vector_129"
+              d="M254.17 215.41C248.87 226.41 211.79 271.59 171.89 260.52C142.61 252.41 122.56 230.92 115.65 224.52C108.74 218.12 100.9 200.89 99.05 189.91C97.2 178.93 102.77 148.05 115.65 145.51C120.49 144.51 140.54 161.74 146.99 171.51C153.44 181.28 188.94 192.99 208.3 183.2C227.66 173.41 243.57 153.6 250.48 155.75C257.39 157.9 263.31 196.5 254.17 215.41Z"
+              fill="#491B2A"
+            />
+            <path
+              id="Vector_130"
+              d="M250.48 155.73C243.57 153.58 227.67 173.4 208.3 183.18C188.93 192.96 153.45 181.27 147 171.5C140.55 161.73 120.5 144.5 115.66 145.5C102.78 148.04 97.2201 178.92 99.0601 189.9C100.9 200.88 108.74 218.06 115.66 224.51C116.44 225.24 117.41 226.18 118.53 227.26C125.05 204.74 150.48 187.94 180.88 187.94C212.35 187.94 238.49 205.94 243.88 229.66C249.19 223.66 252.82 218.34 254.22 215.44C263.31 196.5 257.4 157.88 250.48 155.73Z"
+              fill="#401624"
+            />
+            <path
+              id="Vector_131"
+              d="M161.73 224.77C149.52 213.84 166.83 198.71 174.15 200.6C181.47 202.49 180.57 209.66 177.25 212.88C173.93 216.1 186.78 218.54 184.34 214.2C181.9 209.86 178.8 203.81 186.34 201.92C193.88 200.03 213.34 206.45 206.06 220.81C198.78 235.17 170.6 232.71 161.73 224.77Z"
+              fill="#290C16"
+            />
+            <path
+              id="Vector_132"
+              d="M116.59 232.24C145.26 205.94 177.77 219.24 177.77 219.24C177.77 219.24 212.34 207.13 224.16 211.24C235.98 215.35 246.62 229.24 246.62 229.24L218.55 260.57L176 266.5L116.59 232.24Z"
+              fill="#FF4970"
+            />
+            <path
+              id="Vector_133"
+              d="M138.46 240.73C138.46 230.73 160.84 222.56 188.46 222.56C214.23 222.56 235.46 229.65 238.14 238.76L246.64 229.28C246.64 229.28 236 215.39 224.18 211.28C212.36 207.17 177.79 219.28 177.79 219.28C177.79 219.28 145.28 205.99 116.61 232.28L140.76 246.21C139.344 244.729 138.525 242.778 138.46 240.73Z"
+              fill="#DA1F4C"
+            />
+            <path
+              id="Vector_134"
+              d="M188.44 258.91C175.24 258.91 163.24 257.04 154.31 254L176 266.5L218.56 260.59L225.4 252.95C216.25 256.62 203.08 258.91 188.44 258.91Z"
+              fill="#DA1F4C"
+            />
+            <path
+              id="Vector_135"
+              d="M202.88 225.29C202.88 224.72 201.88 224.29 200.67 224.29C199.46 224.29 198.45 224.75 198.45 225.29C198.45 225.83 199.45 226.29 200.67 226.29C201.89 226.29 202.88 225.86 202.88 225.29Z"
+              fill="#DA1F4C"
+            />
+            <path
+              id="Vector_136"
+              d="M182.64 224.41C182.64 223.83 181.64 223.41 180.43 223.41C179.22 223.41 178.21 223.87 178.21 224.41C178.21 224.95 179.21 225.41 180.43 225.41C181.65 225.41 182.64 224.98 182.64 224.41Z"
+              fill="#DA1F4C"
+            />
+            <path
+              id="Vector_137"
+              d="M176.29 223.22C176.29 222.65 175.29 222.22 174.07 222.22C172.85 222.22 171.86 222.68 171.86 223.22C171.86 223.76 172.86 224.22 174.07 224.22C175.28 224.22 176.29 223.8 176.29 223.22Z"
+              fill="#DA1F4C"
+            />
+            <path
+              id="Vector_138"
+              d="M159.15 227.21C159.15 226.64 158.15 226.21 156.93 226.21C155.71 226.21 154.72 226.67 154.72 227.21C154.72 227.75 155.72 228.21 156.93 228.21C158.14 228.21 159.15 227.78 159.15 227.21Z"
+              fill="#DA1F4C"
+            />
+            <path
+              id="Vector_139"
+              d="M196.83 222.78C196.83 222.21 195.83 221.78 194.61 221.78C193.39 221.78 192.39 222.24 192.39 222.78C192.39 223.32 193.39 223.78 194.61 223.78C195.83 223.78 196.83 223.35 196.83 222.78Z"
+              fill="#DA1F4C"
+            />
+            <path
+              id="Vector_140"
+              d="M189.29 222.93C189.29 222.36 188.29 221.93 187.07 221.93C185.85 221.93 184.86 222.4 184.86 222.93C184.86 223.46 185.86 223.93 187.07 223.93C188.28 223.93 189.29 223.5 189.29 222.93Z"
+              fill="#DA1F4C"
+            />
+            <path
+              id="Vector_141"
+              d="M164.32 225.29C164.32 224.72 163.32 224.29 162.1 224.29C160.88 224.29 159.89 224.75 159.89 225.29C159.89 225.83 160.89 226.29 162.1 226.29C163.31 226.29 164.32 225.86 164.32 225.29Z"
+              fill="#DA1F4C"
+            />
+            <path
+              id="Vector_142"
+              d="M167.2 227.95C167.2 227.63 166.64 227.37 165.95 227.37C165.26 227.37 164.69 227.63 164.69 227.95C164.69 228.27 165.25 228.54 165.95 228.54C166.65 228.54 167.2 228.28 167.2 227.95Z"
+              fill="#DA1F4C"
+            />
+            <path
+              id="Vector_143"
+              d="M172.82 225.59C172.82 225.26 172.25 225 171.56 225C170.87 225 170.3 225.26 170.3 225.59C170.3 225.92 170.87 226.17 171.56 226.17C172.25 226.17 172.82 225.91 172.82 225.59Z"
+              fill="#DA1F4C"
+            />
+            <path
+              id="Vector_144"
+              d="M178.28 228.1C178.28 227.78 177.72 227.51 177.03 227.51C176.34 227.51 175.77 227.78 175.77 228.1C175.77 228.42 176.33 228.69 177.03 228.69C177.73 228.69 178.28 228.42 178.28 228.1Z"
+              fill="#DA1F4C"
+            />
+            <path
+              id="Vector_145"
+              d="M193.8 226.78C193.8 226.45 193.24 226.19 192.54 226.19C191.84 226.19 191.29 226.45 191.29 226.78C191.29 227.11 191.85 227.36 192.54 227.36C193.23 227.36 193.8 227.1 193.8 226.78Z"
+              fill="#DA1F4C"
+            />
+            <path
+              id="Vector_146"
+              d="M186.85 229.43C186.85 229.11 186.29 228.85 185.6 228.85C184.91 228.85 184.34 229.11 184.34 229.43C184.34 229.75 184.9 230.02 185.6 230.02C186.3 230.02 186.85 229.76 186.85 229.43Z"
+              fill="#DA1F4C"
+            />
+            <path
+              id="Vector_147"
+              d="M202.47 229.32C202.56 229.01 202.08 228.61 201.41 228.43C200.74 228.25 200.13 228.35 200.05 228.66C199.97 228.97 200.44 229.38 201.11 229.56C201.78 229.74 202.39 229.63 202.47 229.32Z"
+              fill="#DA1F4C"
+            />
+            <path
+              id="Vector_148"
+              d="M208.7 230.38C208.77 230.06 208.27 229.69 207.59 229.54C206.91 229.39 206.31 229.54 206.24 229.86C206.17 230.18 206.67 230.55 207.35 230.69C208.03 230.83 208.63 230.69 208.7 230.38Z"
+              fill="#DA1F4C"
+            />
+            <path
+              id="Vector_149"
+              d="M159.81 231.2C159.81 230.88 159.25 230.62 158.56 230.62C157.87 230.62 157.3 230.88 157.3 231.2C157.3 231.52 157.86 231.79 158.56 231.79C159.26 231.79 159.81 231.5 159.81 231.2Z"
+              fill="#DA1F4C"
+            />
+            <path
+              id="Vector_150"
+              d="M169.89 223.98C169.78 223.42 168.72 223.15 167.52 223.39C166.32 223.63 165.43 224.28 165.52 224.84C165.61 225.4 166.7 225.66 167.9 225.42C169.1 225.18 170 224.5 169.89 223.98Z"
+              fill="#DA1F4C"
+            />
+            <path
+              id="Vector_151"
+              d="M162.55 216.76C162.55 216.45 161.55 216.21 160.33 216.21C159.11 216.21 158.11 216.45 158.11 216.76C158.11 217.07 159.11 217.31 160.33 217.31C161.55 217.31 162.55 217.06 162.55 216.76Z"
+              fill="#BA1947"
+            />
+            <path
+              id="Vector_152"
+              d="M165.87 220.64C165.87 220.31 165.31 220.05 164.62 220.05C163.93 220.05 163.36 220.31 163.36 220.64C163.36 220.97 163.92 221.22 164.62 221.22C165.32 221.22 165.87 220.96 165.87 220.64Z"
+              fill="#BA1947"
+            />
+            <path
+              id="Vector_153"
+              d="M173.7 219.72C173.7 219.37 173.14 219.09 172.45 219.09C171.76 219.09 171.19 219.37 171.19 219.72C171.19 220.07 171.75 220.35 172.45 220.35C173.15 220.35 173.7 220.07 173.7 219.72Z"
+              fill="#BA1947"
+            />
+            <path
+              id="Vector_154"
+              d="M169.18 217.6C169.18 217.23 168.24 216.85 167.02 216.76C165.8 216.67 164.78 216.9 164.76 217.28C164.74 217.66 165.76 218.03 166.93 218.12C168.1 218.21 169.16 217.98 169.18 217.6Z"
+              fill="#BA1947"
+            />
+            <path
+              id="Vector_155"
+              d="M160.14 220.03C160.14 219.42 159.08 219.03 157.86 219.09C156.64 219.15 155.68 219.73 155.72 220.34C155.76 220.95 156.77 221.34 157.99 221.34C159.21 221.34 160.18 220.64 160.14 220.03Z"
+              fill="#BA1947"
+            />
+            <path
+              id="Vector_156"
+              d="M135 223.78C135 223.21 134 222.78 132.72 222.89C131.44 223 130.54 223.49 130.58 224.06C130.62 224.63 131.64 225.06 132.86 224.95C134.08 224.84 135 224.35 135 223.78Z"
+              fill="#BA1947"
+            />
+            <path
+              id="Vector_157"
+              d="M142.57 221.41C142.57 220.84 141.51 220.41 140.29 220.52C139.07 220.63 138.11 221.13 138.14 221.7C138.17 222.27 139.2 222.7 140.42 222.59C141.64 222.48 142.61 221.98 142.57 221.41Z"
+              fill="#BA1947"
+            />
+            <path
+              id="Vector_158"
+              d="M139.35 233.16C139.35 232.59 138.35 232.16 137.08 232.27C135.81 232.38 134.89 232.88 134.93 233.45C134.97 234.02 135.99 234.45 137.21 234.33C138.43 234.21 139.39 233.73 139.35 233.16Z"
+              fill="#BA1947"
+            />
+            <path
+              id="Vector_159"
+              d="M145.17 224.1C145.17 223.78 144.57 223.55 143.88 223.6C143.19 223.65 142.65 223.94 142.67 224.26C142.69 224.58 143.27 224.81 143.96 224.77C144.65 224.73 145.19 224.42 145.17 224.1Z"
+              fill="#BA1947"
+            />
+            <path
+              id="Vector_160"
+              d="M142.34 227.21C142.34 226.89 141.74 226.66 141.05 226.71C140.36 226.76 139.81 227.05 139.83 227.38C139.85 227.71 140.43 227.92 141.12 227.88C141.81 227.84 142.36 227.5 142.34 227.21Z"
+              fill="#BA1947"
+            />
+            <path
+              id="Vector_161"
+              d="M150.62 221.37C150.62 221.05 150.02 220.83 149.33 220.87C148.64 220.91 148.09 221.22 148.12 221.54C148.15 221.86 148.71 222.09 149.41 222.04C150.11 221.99 150.64 221.7 150.62 221.37Z"
+              fill="#BA1947"
+            />
+            <path
+              id="Vector_162"
+              d="M156.24 223.5C156.24 223.17 155.64 222.95 154.95 222.99C154.26 223.03 153.71 223.34 153.73 223.66C153.75 223.98 154.33 224.21 155.03 224.16C155.73 224.11 156.26 223.85 156.24 223.5Z"
+              fill="#BA1947"
+            />
+            <path
+              id="Vector_163"
+              d="M138 227.82C138 227.5 137.4 227.28 136.71 227.32C136.02 227.36 135.48 227.67 135.5 227.99C135.52 228.31 136.1 228.54 136.79 228.49C137.48 228.44 138 228.15 138 227.82Z"
+              fill="#BA1947"
+            />
+            <path
+              id="Vector_164"
+              d="M149.48 218.29C149.34 217.74 148.26 217.55 147.08 217.87C145.9 218.19 145.08 218.87 145.2 219.44C145.32 220.01 146.43 220.18 147.61 219.86C148.79 219.54 149.63 218.85 149.48 218.29Z"
+              fill="#BA1947"
+            />
+            <path
+              id="Vector_165"
+              d="M211.45 224.85C211.45 224.28 210.45 223.85 209.24 223.85C208.03 223.85 207.02 224.32 207.02 224.85C207.02 225.38 208.02 225.85 209.24 225.85C210.46 225.85 211.45 225.42 211.45 224.85Z"
+              fill="#DA1F4C"
+            />
+            <path
+              id="Vector_166"
+              d="M206.73 223.08C206.73 222.5 205.73 222.08 204.51 222.08C203.29 222.08 202.29 222.54 202.29 223.08C202.29 223.62 203.29 224.08 204.51 224.08C205.73 224.08 206.73 223.65 206.73 223.08Z"
+              fill="#DA1F4C"
+            />
+            <path
+              id="Vector_167"
+              d="M197.3 216.72C197.22 216.21 196.3 215.93 195.18 216.11C194.06 216.29 193.26 216.83 193.34 217.34C193.42 217.85 194.34 218.13 195.47 217.95C196.6 217.77 197.38 217.23 197.3 216.72Z"
+              fill="#BA1947"
+            />
+            <path
+              id="Vector_168"
+              d="M192.18 216.14C192.12 215.89 191.18 215.88 190.12 216.14C189.06 216.4 188.2 216.78 188.26 217.03C188.32 217.28 189.26 217.29 190.32 217.03C191.38 216.77 192.23 216.39 192.18 216.14Z"
+              fill="#BA1947"
+            />
+            <path
+              id="Vector_169"
+              d="M183.69 218.7C183.63 218.45 182.69 218.45 181.64 218.7C180.59 218.95 179.74 219.37 179.8 219.62C179.86 219.87 180.8 219.86 181.85 219.62C182.9 219.38 183.75 218.95 183.69 218.7Z"
+              fill="#BA1947"
+            />
+            <path
+              id="Vector_170"
+              d="M189.74 219.32C189.74 219.03 189.2 218.83 188.58 218.87C187.96 218.91 187.46 219.18 187.48 219.47C187.5 219.76 188.02 219.97 188.65 219.92C189.28 219.87 189.76 219.61 189.74 219.32Z"
+              fill="#BA1947"
+            />
+            <path
+              id="Vector_171"
+              d="M183.91 221.14C183.86 220.85 183.32 220.69 182.7 220.79C182.08 220.89 181.62 221.2 181.7 221.49C181.78 221.78 182.28 221.93 182.9 221.84C183.52 221.75 184 221.42 183.91 221.14Z"
+              fill="#BA1947"
+            />
+            <path
+              id="Vector_172"
+              d="M200 219.9C200 219.61 199.55 219.32 198.92 219.25C198.29 219.18 197.77 219.37 197.74 219.66C197.71 219.95 198.19 220.24 198.81 220.31C199.43 220.38 199.94 220.19 200 219.9Z"
+              fill="#BA1947"
+            />
+            <path
+              id="Vector_173"
+              d="M201.64 213.42C201.56 212.91 200.64 212.63 199.51 212.81C198.38 212.99 197.6 213.54 197.68 214.05C197.76 214.56 198.68 214.83 199.8 214.66C200.92 214.49 201.72 213.93 201.64 213.42Z"
+              fill="#BA1947"
+            />
+            <path
+              id="Vector_174"
+              d="M214.74 216.5C214.74 215.93 213.74 215.5 212.53 215.5C211.32 215.5 210.31 215.97 210.31 216.5C210.31 217.03 211.31 217.5 212.53 217.5C213.75 217.5 214.74 217.11 214.74 216.5Z"
+              fill="#BA1947"
+            />
+            <path
+              id="Vector_175"
+              d="M206.47 215.06C207.696 215.06 208.69 214.599 208.69 214.03C208.69 213.461 207.696 213 206.47 213C205.244 213 204.25 213.461 204.25 214.03C204.25 214.599 205.244 215.06 206.47 215.06Z"
+              fill="#BA1947"
+            />
+            <path
+              id="Vector_176"
+              d="M205.65 218.02C205.65 217.7 205.09 217.43 204.4 217.43C203.71 217.43 203.14 217.7 203.14 218.02C203.14 218.34 203.71 218.61 204.4 218.61C205.09 218.61 205.65 218.34 205.65 218.02Z"
+              fill="#BA1947"
+            />
+            <path
+              id="Vector_177"
+              d="M214.33 220.56C214.41 220.25 213.94 219.85 213.27 219.67C212.6 219.49 211.99 219.6 211.9 219.91C211.81 220.22 212.29 220.62 212.96 220.8C213.63 220.98 214.24 220.88 214.33 220.56Z"
+              fill="#BA1947"
+            />
+            <path
+              id="Vector_178"
+              d="M220.56 221.62C220.63 221.3 220.13 220.93 219.45 220.79C218.77 220.65 218.17 220.79 218.1 221.1C218.03 221.41 218.53 221.79 219.21 221.94C219.89 222.09 220.49 221.94 220.56 221.62Z"
+              fill="#BA1947"
+            />
+            <path
+              id="Vector_179"
+              d="M221.1 217.12C222.326 217.12 223.32 216.659 223.32 216.09C223.32 215.521 222.326 215.06 221.1 215.06C219.874 215.06 218.88 215.521 218.88 216.09C218.88 216.659 219.874 217.12 221.1 217.12Z"
+              fill="#BA1947"
+            />
+            <path
+              id="Vector_180"
+              d="M220.36 212.5C220.36 211.93 219.36 211.5 218.14 211.5C216.92 211.5 215.92 211.96 215.92 212.5C215.92 213.04 216.92 213.5 218.14 213.5C219.36 213.5 220.36 213.05 220.36 212.5Z"
+              fill="#BA1947"
+            />
+            <path
+              id="Vector_181"
+              d="M230.34 220.78C230.51 220.23 229.69 219.51 228.52 219.16C227.35 218.81 226.26 218.97 226.1 219.52C225.94 220.07 226.75 220.79 227.93 221.14C229.11 221.49 230.18 221.33 230.34 220.78Z"
+              fill="#BA1947"
+            />
+            <path
+              id="Vector_182"
+              d="M227.25 224.63C227.42 224.36 227.08 223.84 226.49 223.48C225.9 223.12 225.28 223.04 225.11 223.31C224.94 223.58 225.28 224.11 225.87 224.47C226.46 224.83 227.08 224.91 227.25 224.63Z"
+              fill="#BA1947"
+            />
+            <path
+              id="Vector_183"
+              d="M232.92 227.42C233.07 227.14 232.7 226.64 232.09 226.31C231.48 225.98 230.86 225.94 230.71 226.23C230.56 226.52 230.92 227.01 231.53 227.34C232.14 227.67 232.76 227.71 232.92 227.42Z"
+              fill="#BA1947"
+            />
+            <path
+              id="Vector_184"
+              d="M237.13 222.91C237.3 222.37 236.48 221.64 235.3 221.29C234.12 220.94 233.05 221.1 232.89 221.65C232.73 222.2 233.54 222.92 234.71 223.27C235.88 223.62 237 223.5 237.13 222.91Z"
+              fill="#BA1947"
+            />
+            <path
+              id="Vector_185"
+              d="M231.56 217.09C231.72 216.55 230.9 215.82 229.73 215.47C228.56 215.12 227.47 215.28 227.31 215.83C227.15 216.38 227.96 217.1 229.14 217.45C230.32 217.8 231.39 217.64 231.56 217.09Z"
+              fill="#BA1947"
+            />
+            <path
+              id="Vector_186"
+              d="M121 151.19L115.24 146.76L96.18 161.17L91 183.11L95.24 205.5L111.24 212.82C99.72 193.31 99.72 178.46 105.24 180.46C110.76 182.46 103.46 189.55 109.45 188.88C115.44 188.21 112.11 182.67 108.78 174.03C105.45 165.39 109.56 155.75 112 159.74C114.44 163.73 110.45 169.27 115.32 172.37C120.19 175.47 119.87 172.04 118.98 165.83C118.09 159.62 121 151.19 121 151.19Z"
+              fill="#00868F"
+            />
+            <path
+              id="Vector_187"
+              d="M107.77 187.04C107.771 187.64 107.825 188.239 107.93 188.83C108.426 188.915 108.934 188.915 109.43 188.83C113.81 188.34 113.2 185.23 111.25 180.21C109.23 181.18 107.77 183.87 107.77 187.04Z"
+              fill="#00ADB3"
+            />
+            <path
+              id="Vector_188"
+              d="M112.76 165.67C112.62 168.19 112.61 170.67 115.3 172.36C115.782 172.672 116.286 172.946 116.81 173.18C117.08 171.833 116.926 170.436 116.37 169.18C115.48 166.89 113.89 165.37 112.76 165.67Z"
+              fill="#00ADB3"
+            />
+            <path
+              id="Vector_189"
+              d="M115.08 188.76C118.85 192.31 121.29 196.85 118.4 196.96C115.51 197.07 115.08 188.76 115.08 188.76Z"
+              fill="#00ADB3"
+            />
+            <path
+              id="Vector_190"
+              d="M130.19 165.8C131.48 174.96 131.24 180.38 129.09 179.9C124.42 178.85 130.19 165.8 130.19 165.8Z"
+              fill="#00ADB3"
+            />
+            <path
+              id="Vector_191"
+              d="M106.33 185.25C106.311 184.998 106.218 184.758 106.064 184.557C105.91 184.357 105.701 184.206 105.462 184.123C105.224 184.04 104.966 184.028 104.721 184.089C104.475 184.149 104.253 184.281 104.081 184.466C103.909 184.651 103.795 184.882 103.752 185.131C103.71 185.38 103.741 185.636 103.841 185.868C103.942 186.1 104.108 186.297 104.319 186.436C104.53 186.575 104.777 186.65 105.03 186.65C105.207 186.644 105.382 186.603 105.543 186.529C105.704 186.455 105.85 186.35 105.97 186.22C106.091 186.09 106.185 185.937 106.247 185.771C106.308 185.604 106.337 185.428 106.33 185.25Z"
+              fill="#00868F"
+            />
+            <path
+              id="Vector_192"
+              d="M109.8 178.5C107.73 179.68 106.99 178.72 106.55 175.91C106.286 174.165 106.361 172.386 106.77 170.67L109.8 178.5Z"
+              fill="#00ADB3"
+            />
+            <path
+              id="Vector_193"
+              d="M207.72 192.31C206.39 197.4 203.96 201.31 206.72 200.95C209.48 200.59 207.72 192.31 207.72 192.31Z"
+              fill="#00ADB3"
+            />
+            <path
+              id="Vector_194"
+              d="M200.3 192.86C205.06 186.43 207.17 185.32 208.72 188.21C210.27 191.1 208.5 197.07 210.49 195.41C212.48 193.75 211.71 190.41 210.72 187.41C209.73 184.41 213.26 183.75 213.26 183.75L213.6 180.1L204.07 182.2L197.42 188.74L200.3 192.86Z"
+              fill="#00868F"
+            />
+            <path
+              id="Vector_195"
+              d="M236.43 168.26C243.19 170.14 244.96 167.59 244.74 172.47C244.52 177.35 241.52 183.55 245.07 183C248.62 182.45 245.07 172.91 245.73 169.92C246.39 166.93 247.73 165.04 249.28 168.04C250.83 171.04 251.05 172.14 251.94 177.68C252.83 183.22 248.73 187.87 251.83 189.87C254.93 191.87 254.71 188.54 255.26 185.99C255.81 183.44 255.82 167.93 255.82 167.93L253.27 162.16L247.51 160.06L242.51 160.83L236.43 168.26Z"
+              fill="#00868F"
+            />
+            <path
+              id="Vector_196"
+              d="M244.47 175.65H244.31C243.54 179.65 242.31 183.43 245.07 183C245.348 182.959 245.61 182.845 245.83 182.669C246.05 182.493 246.219 182.263 246.32 182C246.54 178.55 245.8 175.85 244.47 175.65Z"
+              fill="#00ADB3"
+            />
+            <path
+              id="Vector_197"
+              d="M251.08 185C250.54 187.15 250.25 188.85 251.83 189.87C252.143 190.075 252.478 190.243 252.83 190.37C252.83 190.31 252.83 190.26 252.83 190.2C252.83 187.9 252.11 185.92 251.08 185Z"
+              fill="#00ADB3"
+            />
+            <path
+              id="Vector_198"
+              d="M253.82 204.83C248.62 202.5 247.06 202.28 248.39 198.07C249.72 193.86 246.29 189.87 246.18 192.42C246.07 194.97 249.06 200.28 246.84 203.72C244.62 207.16 248.39 214.36 248.39 214.36L253.05 211.92L253.82 204.83Z"
+              fill="#00868F"
+            />
+            <path
+              id="Vector_199"
+              d="M120.51 148.75C115.41 156.51 114.75 165.82 117.41 169.59C120.07 173.36 121.84 174.02 125.16 173.36C128.48 172.7 132.26 168.26 132.26 168.26L134.47 156.26L120.51 148.75Z"
+              fill="#F2E5B2"
+            />
+            <path
+              id="Vector_200"
+              d="M120.51 148.75C118.011 152.554 116.464 156.903 116 161.43C120.09 152.23 128.66 159.57 129.4 160.72C130.14 161.87 129.65 166.64 127.62 172.32C129.349 171.182 130.916 169.817 132.28 168.26L134.49 156.26L120.51 148.75Z"
+              fill="#E5C699"
+            />
+            <path
+              id="Vector_201"
+              d="M242.63 229.21C240.41 217.21 244.4 204.61 246.63 203.94C248.86 203.27 252.71 211.05 253.5 212.14C255.27 214.61 244.1 230.29 242.63 229.21Z"
+              fill="#F2E5B2"
+            />
+            <path
+              id="Vector_202"
+              d="M232.44 237.41C230.22 234.08 229.55 229.87 231.55 228.99C233.55 228.11 237.09 229.21 237.09 229.21C237.09 229.21 240.25 226.26 242.19 228.1C245.51 231.25 234.39 239.43 232.44 237.41Z"
+              fill="#F2E5B2"
+            />
+            <path
+              id="Vector_203"
+              d="M241.84 227.84C240.69 229.45 238.95 230.21 237.72 229.54C237.523 229.427 237.348 229.282 237.2 229.11L237.09 229.21H236.93C236.93 229.21 236.93 229.27 236.93 229.3C236.93 230.87 235.05 232.14 232.72 232.14C231.907 232.148 231.102 231.981 230.36 231.65C230.514 233.738 231.23 235.745 232.43 237.46C234.38 239.46 245.5 231.3 242.18 228.15C242.078 228.034 241.964 227.93 241.84 227.84Z"
+              fill="#E5C699"
+            />
+            <path
+              id="Vector_204"
+              d="M216.26 249.38C215.37 238.74 208.94 220.12 209.39 218.38C209.84 216.64 228.89 224.58 233.1 236.77C237.31 248.96 216.26 249.38 216.26 249.38Z"
+              fill="#F2E5B2"
+            />
+            <path
+              id="Vector_205"
+              d="M201 233.42C200.34 233.86 194.8 246.5 199.9 256.69C205.9 255.36 212.17 254.23 215.41 250.04C213.38 240.07 201.63 232.98 201 233.42Z"
+              fill="#F2E5B2"
+            />
+            <path
+              id="Vector_206"
+              d="M168.61 256.5C170.38 247.86 175.48 238.77 177.47 238.99C179.46 239.21 183.23 256.28 183.23 258.99C175.92 262.67 168.61 256.5 168.61 256.5Z"
+              fill="#F2E5B2"
+            />
+            <path
+              id="Vector_207"
+              d="M154.42 252.7C156.54 249.81 160.08 248.09 162.91 248.45C165.74 248.81 165.5 251.94 165.5 251.94C165.5 251.94 168.57 252.71 167.86 256.33C167.15 259.95 154.42 252.7 154.42 252.7Z"
+              fill="#F2E5B2"
+            />
+            <path
+              id="Vector_208"
+              d="M154.2 218.13C153.31 217.9 139.13 222.78 135.14 239.85L153.76 252.26C157.52 240.73 155.09 218.35 154.2 218.13Z"
+              fill="#F2E5B2"
+            />
+            <path
+              id="Vector_209"
+              d="M117.26 225.96C120.47 223.05 130.09 223.18 132.16 224.57C134.23 225.96 135.37 229.64 135.16 232.76C134.869 234.959 134.086 237.065 132.87 238.92C125.21 235.41 119.64 231.21 117.26 225.96Z"
+              fill="#F2E5B2"
+            />
+            <path
+              id="Vector_210"
+              d="M116.52 224.85C120.07 222.19 117.19 219.97 117.19 219.97C117.19 219.97 122.95 215.32 121.19 210.89C119.43 206.46 116.09 204.46 114.76 203.13C113.43 201.8 114.76 199.13 113.87 198.92C112.98 198.71 104.78 197.14 103.87 202.69C102.96 208.24 113 222.63 113 222.63L116.52 224.85Z"
+              fill="#F2E5B2"
+            />
+            <path
+              id="Vector_211"
+              d="M248.17 157.62C248.84 163.38 246.4 166.04 244.85 167.37C243.3 168.7 237.76 168.04 236.2 166.71C235.733 166.251 235.372 165.695 235.142 165.082C234.912 164.469 234.819 163.813 234.87 163.16L244.63 155.63L248.17 157.62Z"
+              fill="#F2E5B2"
+            />
+            <path
+              id="Vector_212"
+              d="M230.88 166.5C231.55 169.15 234.21 174.92 235.32 174.7C236.43 174.48 238.2 166.5 238.2 164.94C238.2 163.38 235.54 160.29 235.54 160.29L230.88 166.5Z"
+              fill="#F2E5B2"
+            />
+            <path
+              id="Vector_213"
+              d="M209.39 179.78C211.6 190.2 218.25 198.4 223.13 202.61C223.13 202.61 235.32 183.11 232.44 166.26L209.39 179.78Z"
+              fill="#F2E5B2"
+            />
+            <path
+              id="Vector_214"
+              d="M188.33 186.65C188.55 191.31 194.54 196.85 199.86 196.65C199.86 196.65 205.17 188.89 205.86 187.34C206.361 185.607 206.66 183.822 206.75 182.02L188.33 186.65Z"
+              fill="#F2E5B2"
+            />
+            <path
+              id="Vector_215"
+              d="M170.82 183.77C169.49 188.87 170.38 192.86 172.82 196.18C174.259 198.162 176.161 199.762 178.36 200.84L181.68 197.74L185.23 198.4C185.23 198.4 189.88 189.09 190.33 186.65C190.78 184.21 170.82 183.77 170.82 183.77Z"
+              fill="#F2E5B2"
+            />
+            <path
+              id="Vector_216"
+              d="M158 180.01C157.199 184.244 156.9 188.557 157.11 192.86C157.11 192.86 165.98 193.97 168.41 191.75C170.84 189.53 172.18 185.1 172.18 185.1L158 180.01Z"
+              fill="#F2E5B2"
+            />
+            <path
+              id="Vector_217"
+              d="M141.34 163.16C132.92 179.16 140.01 199.51 144.45 203.28C154.64 202.83 160.63 179.34 160.63 179.34L146.44 165.82L141.34 163.16Z"
+              fill="#F2E5B2"
+            />
+            <path
+              id="Vector_218"
+              d="M233.1 236.74C231.898 233.576 230.009 230.717 227.57 228.37C228.91 233.89 227.99 238.63 224.9 240.05C221.6 241.55 216.9 238.81 213.16 233.71C214.591 238.847 215.628 244.085 216.26 249.38C216.26 249.38 237.31 248.93 233.1 236.74Z"
+              fill="#E5C699"
+            />
+            <path
+              id="Vector_219"
+              d="M210.91 241.23C211.86 244.89 211.09 248.42 208.74 249.9C205.74 251.8 201.26 249.61 198.74 245.01C198.435 244.433 198.165 243.839 197.93 243.23C197.224 247.806 197.885 252.488 199.83 256.69C205.83 255.36 212.1 254.23 215.34 250.04C214.601 246.787 213.08 243.763 210.91 241.23Z"
+              fill="#E5C699"
+            />
+            <path
+              id="Vector_220"
+              d="M183.9 258.69C182.35 253.81 183.45 249.16 185.67 247.83C187.89 246.5 196.75 247.16 198.75 248.49C200.75 249.82 202.07 257.36 202.07 257.36C196.277 259.588 189.955 260.05 183.9 258.69Z"
+              fill="#F2E5B2"
+            />
+            <path
+              id="Vector_221"
+              d="M198.75 248.5C198.181 248.175 197.563 247.946 196.92 247.82C196.975 248.325 196.975 248.835 196.92 249.34C196.771 250.877 196.154 252.331 195.151 253.506C194.149 254.681 192.811 255.519 191.316 255.909C189.822 256.298 188.244 256.219 186.796 255.683C185.348 255.147 184.1 254.179 183.22 252.91C183.049 254.865 183.274 256.834 183.88 258.7C189.935 260.061 196.257 259.598 202.05 257.37C202.05 257.37 200.74 249.82 198.75 248.5Z"
+              fill="#E5C699"
+            />
+            <path
+              id="Vector_222"
+              d="M197.83 244.61C197.734 245.744 197.704 246.883 197.74 248.02C199.567 251.018 200.38 254.524 200.06 258.02C200.72 257.82 201.39 257.61 202.06 257.37C202.06 257.37 201.98 256.93 201.84 256.26L203.45 255.9C200.25 252.66 198.61 247.82 197.83 244.61Z"
+              fill="#D8A96B"
+            />
+            <path
+              id="Vector_223"
+              d="M253.49 212.14C253.11 211.61 251.98 209.47 250.68 207.51C250.55 208.46 250.36 209.43 250.12 210.43C248.56 216.64 245.28 221.19 242.78 220.59C242.5 220.513 242.238 220.381 242.01 220.2C241.892 223.214 242.1 226.231 242.63 229.2C244.1 230.29 255.27 214.61 253.49 212.14Z"
+              fill="#E5C699"
+            />
+            <path
+              id="Vector_224"
+              d="M230.52 230.28C230.396 230.722 230.345 231.182 230.37 231.64L230.05 231.22C231.374 235.35 232.072 239.654 232.12 243.99C232.832 243.131 233.32 242.108 233.54 241.014C233.76 239.92 233.705 238.788 233.38 237.72C233.63 237.714 233.878 237.68 234.12 237.62C232.5 235.403 231.282 232.918 230.52 230.28Z"
+              fill="#D8A96B"
+            />
+            <path
+              id="Vector_225"
+              d="M181.78 250.79C181.09 253.69 179.78 255.65 178.01 255.86C175.44 256.17 172.81 252.56 171.53 247.38C170.25 250.303 169.272 253.348 168.61 256.47C168.61 256.47 175.92 262.67 183.23 258.91C182.946 256.171 182.461 253.458 181.78 250.79Z"
+              fill="#E5C699"
+            />
+            <path
+              id="Vector_226"
+              d="M165.37 254.73C164.74 254.14 164.81 252.83 165.5 251.79C165.51 251.485 165.483 251.179 165.42 250.88C164.933 251.534 164.316 252.079 163.606 252.48C162.897 252.882 162.111 253.13 161.3 253.21C158.92 253.47 156.96 252.31 156.55 250.54C155.743 251.157 155.026 251.884 154.42 252.7C154.42 252.7 167.15 259.96 167.86 256.33C168.039 255.568 168.001 254.771 167.75 254.03C167 254.99 166 255.31 165.37 254.73Z"
+              fill="#E5C699"
+            />
+            <path
+              id="Vector_227"
+              d="M149.18 245.31C143.99 245.31 139.77 238.15 139.66 229.25C137.524 232.481 135.993 236.072 135.14 239.85L153.76 252.26C154.83 248.585 155.454 244.794 155.62 240.97C153.92 243.66 151.66 245.31 149.18 245.31Z"
+              fill="#E5C699"
+            />
+            <path
+              id="Vector_228"
+              d="M133.42 233.82C128.91 235.82 123.66 233.38 121.7 228.43C121.175 227.098 120.93 225.672 120.98 224.24C119.621 224.509 118.345 225.099 117.26 225.96C119.64 231.21 125.26 235.41 132.85 238.96C134.052 237.117 134.83 235.03 135.13 232.85C134.601 233.242 134.027 233.567 133.42 233.82Z"
+              fill="#E5C699"
+            />
+            <path
+              id="Vector_229"
+              d="M117.19 219.97C117.19 219.97 117.59 219.64 118.19 219.09C115.42 218.38 112.84 214.55 112.04 209.61C111.33 205.18 112.26 201.15 114.19 199.25C114.178 199.189 114.153 199.132 114.117 199.082C114.08 199.032 114.034 198.99 113.98 198.96C113.86 199.19 113.74 199.42 113.6 199.65C112.27 201.75 110.1 202.76 108.75 201.91C108.311 201.546 107.989 201.06 107.825 200.514C107.662 199.968 107.663 199.385 107.83 198.84C106.855 198.983 105.951 199.433 105.249 200.126C104.548 200.818 104.086 201.717 103.93 202.69C103.04 208.23 113.01 222.63 113.01 222.63L116.56 224.85C120.07 222.19 117.19 219.97 117.19 219.97Z"
+              fill="#E5C699"
+            />
+            <path
+              id="Vector_230"
+              d="M147.11 170.76C152.54 171.17 156.25 180.54 155.39 191.68C155.343 192.28 155.283 192.87 155.21 193.45C157.496 188.945 159.312 184.217 160.63 179.34L146.44 165.82L141.34 163.16C138.511 168.763 137.168 174.999 137.44 181.27C139.56 174.76 143.25 170.5 147.11 170.76Z"
+              fill="#E5C699"
+            />
+            <path
+              id="Vector_231"
+              d="M157.33 192.89C158.33 188.38 159.5 185.97 160.85 185.62C162.55 185.18 168.97 184.14 169.49 189.9C169.49 190.1 169.49 190.27 169.49 190.45C170.623 188.806 171.506 187.003 172.11 185.1L158 180.01C157.199 184.244 156.9 188.557 157.11 192.86L157.33 192.89Z"
+              fill="#E5C699"
+            />
+            <path
+              id="Vector_232"
+              d="M160.21 180.81C160.49 179.89 160.63 179.34 160.63 179.34L156 174.9C156.52 177.84 160.16 179.9 156.38 191.16C156.61 190.67 156.83 190.16 157.04 189.68C157.04 191.68 157.11 192.86 157.11 192.86H157.18C157.65 186.96 159.4 183.13 161.03 181.09L160.21 180.81Z"
+              fill="#D8A96B"
+            />
+            <path
+              id="Vector_233"
+              d="M172.45 186.36C175.45 185.36 180.5 188.72 181.16 192.49C181.57 194.77 180.7 197.49 179.95 199.35L181.68 197.74L183.37 198.05C182.55 195.39 181.77 191.97 182.57 190.5C183.97 187.91 187.37 190.64 187.37 190.64L185.88 197.07C187.18 194.41 189.99 188.52 190.33 186.65C190.77 184.22 170.82 183.77 170.82 183.77C169.921 186.776 169.998 189.99 171.04 192.95C170.34 189.29 171 186.82 172.45 186.36Z"
+              fill="#E5C699"
+            />
+            <path
+              id="Vector_234"
+              d="M188.33 186.65C188.47 189.58 190.89 192.86 193.99 194.83C189.99 191.41 192.04 188.73 194.83 188.57C198.62 188.36 205.46 183.99 205.56 187.87C205.68 187.65 205.78 187.45 205.84 187.32C206.341 185.587 206.64 183.802 206.73 182L188.33 186.65Z"
+              fill="#E5C699"
+            />
+            <path
+              id="Vector_235"
+              d="M214.63 182.37C216.85 178.6 228.52 171.37 230 179.37C230.38 181.834 230.468 184.335 230.26 186.82C232.26 180.55 233.62 173.2 232.44 166.29L209.39 179.81C211.44 189.49 217.33 197.25 222.06 201.68C213.34 190.14 212.64 185.75 214.63 182.37Z"
+              fill="#E5C699"
+            />
+            <path
+              id="Vector_236"
+              d="M234.36 167.08C235.59 166.33 236.63 168.2 237.19 170.26C237.673 168.515 238.008 166.732 238.19 164.93C238.19 163.38 235.53 160.28 235.53 160.28L230.88 166.5C231.526 168.711 232.429 170.839 233.57 172.84C232.67 170.01 233.13 167.82 234.36 167.08Z"
+              fill="#E5C699"
+            />
+            <path
+              id="Vector_237"
+              d="M240.86 163.31C242.5 161.6 247.23 161.31 247.39 164.05C248.234 162.019 248.504 159.794 248.17 157.62L244.63 155.62L234.87 163.15C234.819 163.803 234.912 164.459 235.142 165.072C235.372 165.686 235.733 166.241 236.2 166.7C237.099 167.304 238.126 167.691 239.2 167.83C238.55 166.25 239.67 164.5 240.86 163.31Z"
+              fill="#E5C699"
+            />
+            <path
+              id="Vector_238"
+              d="M132.7 157.18C129.38 164.27 129.38 170.25 130.26 171.58C131.14 172.91 131.81 174.24 133.81 174.46C135.81 174.68 140.01 172.91 140.01 172.91L142.45 164.49L132.7 157.18Z"
+              fill="#F2E5B2"
+            />
+            <path
+              id="Vector_239"
+              d="M132.7 157.18C129.48 164.04 129.38 169.85 130.18 171.43L130.85 165.89C132.598 164.742 134.674 164.197 136.76 164.34C139.94 164.71 139.76 168.41 139.2 169.88C138.78 171.08 138.56 172.18 133.75 174.45H133.81C135.81 174.68 140.01 172.9 140.01 172.9L142.45 164.48L132.7 157.18Z"
+              fill="#E5C699"
+            />
+            <path
+              id="Vector_240"
+              d="M129.66 170.81C130.11 165.63 130.95 164.31 131.52 163.46L134.7 158.67L134.11 158.24L134.47 156.24L131.92 154.86C131.52 158.05 130.92 162.99 130.63 163.92C130.34 164.85 129.33 168.84 128.51 171.66C128.92 171.42 129.3 171.12 129.66 170.81Z"
+              fill="#D8A96B"
+            />
+            <path
+              id="Vector_241"
+              d="M144.15 164.62L141.75 163.37C141.94 167.11 140.07 170.94 137.68 173.76L137.92 173.69C137.82 174.31 137.74 174.94 137.67 175.56C139.47 171.67 141.81 167.23 144.15 164.62Z"
+              fill="#D8A96B"
+            />
+            <path
+              id="Vector_242"
+              d="M170.82 183.77C170.75 184.04 170.69 184.3 170.64 184.56L170.15 184.38C170.142 186.392 169.962 188.399 169.61 190.38C169.81 190.11 169.99 189.82 170.17 189.54C170.628 187.576 171.299 185.668 172.17 183.85L170.82 183.77Z"
+              fill="#D8A96B"
+            />
+            <path
+              id="Vector_243"
+              d="M190.15 186.19C189.401 185.661 188.537 185.318 187.63 185.19C188.252 187.003 188.642 188.888 188.79 190.8C188.95 190.43 189.1 190.08 189.24 189.73C189.507 190.247 189.815 190.742 190.16 191.21C190.234 189.388 190.609 187.59 191.27 185.89L190.15 186.19Z"
+              fill="#D8A96B"
+            />
+            <path
+              id="Vector_244"
+              d="M230.88 166.5L231.05 167.09L230.37 167.49C231.27 168.18 232.17 170.37 232.76 174.49C232.83 173.49 232.87 172.49 232.87 171.49C232.903 171.562 232.94 171.632 232.98 171.7C232.481 169.534 232.11 167.34 231.87 165.13L230.88 166.5Z"
+              fill="#D8A96B"
+            />
+            <path
+              id="Vector_245"
+              d="M236.68 161.76C236.09 160.92 235.54 160.28 235.54 160.28L235.19 160.74C236.19 162.2 237.19 164.29 237.66 168.27C237.72 168.01 237.77 167.75 237.82 167.5L238.39 167.66C237.797 165.66 237.306 163.631 236.92 161.58L236.68 161.76Z"
+              fill="#D8A96B"
+            />
+            <path
+              id="Vector_246"
+              d="M224.79 183.55C223.01 183.55 221.58 188.19 221.58 193.91C221.562 196.032 221.784 198.148 222.24 200.22L222.68 200.62C224.645 197.318 226.357 193.871 227.8 190.31C227.35 186.37 226.17 183.55 224.79 183.55Z"
+              fill="#F4EBCD"
+            />
+            <path
+              id="Vector_247"
+              d="M218.55 232.34C220.35 231.02 219.4 226.68 216.44 222.64C216.26 222.39 216.07 222.15 215.89 221.92C213.29 220.76 211.39 220.34 211.27 220.85C211.08 221.61 212.35 225.97 213.78 231.29C215.69 232.65 217.48 233.12 218.55 232.34Z"
+              fill="#F4EBCD"
+            />
+            <path
+              id="Vector_248"
+              d="M143.32 188.38C142.78 188.512 142.279 188.772 141.86 189.137C141.44 189.503 141.115 189.963 140.91 190.48C141.91 194.88 143.57 198.3 144.91 199.48C146.7 199.41 148.29 198.28 149.68 196.65C148.39 191.5 145.76 188.01 143.32 188.38Z"
+              fill="#F4EBCD"
+            />
+            <path
+              id="Vector_249"
+              d="M210.34 250.65C210.183 250.43 209.948 250.277 209.683 250.222C209.419 250.166 209.143 250.212 208.91 250.35C208.31 250.61 207.97 251.17 208.15 251.6C208.307 251.82 208.542 251.973 208.807 252.029C209.071 252.084 209.347 252.038 209.58 251.9C210.18 251.64 210.52 251.08 210.34 250.65Z"
+              fill="#D8A96B"
+            />
+            <path
+              id="Vector_250"
+              d="M209.46 247.35C209.402 247.214 209.302 247.1 209.175 247.024C209.048 246.949 208.9 246.915 208.752 246.929C208.605 246.942 208.466 247.002 208.355 247.1C208.243 247.197 208.166 247.328 208.134 247.472C208.101 247.617 208.115 247.767 208.174 247.903C208.232 248.039 208.333 248.153 208.46 248.228C208.587 248.304 208.735 248.337 208.883 248.323C209.03 248.309 209.169 248.248 209.28 248.15C209.56 247.9 209.64 247.55 209.46 247.35Z"
+              fill="#E5C699"
+            />
+            <path
+              id="Vector_251"
+              d="M209.47 242.12C209.32 242.12 209.2 242.29 209.21 242.5C209.22 242.71 209.33 242.88 209.48 242.88C209.63 242.88 209.75 242.7 209.74 242.49C209.73 242.28 209.62 242.12 209.47 242.12Z"
+              fill="#E5C699"
+            />
+            <path
+              id="Vector_252"
+              d="M205.59 250.17C205.59 249.67 205.04 249.24 204.33 249.22C203.62 249.2 203.04 249.59 203.03 250.09C203.02 250.59 203.57 251.01 204.28 251.03C204.99 251.05 205.57 250.67 205.59 250.17Z"
+              fill="#E5C699"
+            />
+            <path
+              id="Vector_253"
+              d="M192.71 255.76C192.554 255.538 192.32 255.384 192.054 255.328C191.789 255.272 191.512 255.319 191.28 255.46C190.68 255.72 190.34 256.28 190.53 256.71C190.684 256.931 190.917 257.085 191.181 257.141C191.444 257.197 191.72 257.15 191.95 257.01C192.56 256.75 192.89 256.19 192.71 255.76Z"
+              fill="#D8A96B"
+            />
+            <path
+              id="Vector_254"
+              d="M174.26 256.38C174.31 256.09 174.03 255.79 173.62 255.71C173.21 255.63 172.83 255.8 172.77 256.08C172.71 256.36 173 256.68 173.41 256.76C173.82 256.84 174.2 256.67 174.26 256.38Z"
+              fill="#D8A96B"
+            />
+            <path
+              id="Vector_255"
+              d="M172.5 253.5C172.599 253.367 172.654 253.206 172.658 253.04C172.662 252.873 172.613 252.71 172.52 252.573C172.426 252.435 172.292 252.33 172.136 252.273C171.98 252.215 171.81 252.207 171.65 252.25C171.4 252.42 171.21 253.16 171.65 253.18C172.09 253.2 172.25 253.7 172.5 253.5Z"
+              fill="#D8A96B"
+            />
+            <path
+              id="Vector_256"
+              d="M198.56 256.66C198.25 255.94 197.17 255.72 196.15 256.16C195.13 256.6 194.57 257.54 194.88 258.26C195.141 258.633 195.535 258.892 195.981 258.985C196.427 259.078 196.891 258.998 197.28 258.76C198.3 258.32 198.87 257.38 198.56 256.66Z"
+              fill="#D8A96B"
+            />
+            <path
+              id="Vector_257"
+              d="M180.55 258.37C180.55 257.59 179.67 256.94 178.55 256.93C177.43 256.92 176.55 257.54 176.55 258.32C176.55 259.1 177.43 259.75 178.55 259.76C179.67 259.77 180.54 259.15 180.55 258.37Z"
+              fill="#D8A96B"
+            />
+            <path
+              id="Vector_258"
+              d="M140.36 241.83C140.11 240.83 139.13 240.83 138.44 241.48C138.204 241.699 138.025 241.971 137.917 242.274C137.809 242.577 137.776 242.902 137.82 243.22C137.843 243.352 137.892 243.479 137.964 243.592C138.037 243.705 138.132 243.802 138.243 243.878C138.354 243.953 138.479 244.006 138.611 244.032C138.743 244.058 138.878 244.058 139.01 244.03C139.475 243.907 139.874 243.61 140.125 243.2C140.377 242.791 140.461 242.3 140.36 241.83Z"
+              fill="#D8A96B"
+            />
+            <path
+              id="Vector_259"
+              d="M106.91 204.66C105.99 204.28 105.39 205.05 105.49 205.99C105.523 206.306 105.631 206.61 105.805 206.876C105.979 207.142 106.214 207.363 106.49 207.52C106.61 207.583 106.742 207.62 106.877 207.63C107.012 207.64 107.148 207.623 107.276 207.579C107.404 207.535 107.522 207.465 107.622 207.374C107.723 207.284 107.804 207.173 107.86 207.05C108.035 206.606 108.034 206.111 107.858 205.667C107.681 205.223 107.342 204.863 106.91 204.66Z"
+              fill="#D8A96B"
+            />
+            <path
+              id="Vector_260"
+              d="M191.84 252.5C191.711 252.429 191.564 252.4 191.418 252.417C191.272 252.433 191.135 252.494 191.025 252.592C190.916 252.689 190.839 252.819 190.806 252.962C190.773 253.105 190.784 253.254 190.84 253.39C190.964 253.478 191.116 253.518 191.267 253.501C191.418 253.485 191.558 253.413 191.66 253.3C191.784 253.213 191.871 253.083 191.904 252.935C191.938 252.787 191.915 252.632 191.84 252.5Z"
+              fill="#E5C699"
+            />
+            <path
+              id="Vector_261"
+              d="M185.8 252.26C185.8 252 185.52 251.78 185.15 251.77C184.78 251.76 184.48 251.96 184.47 252.22C184.46 252.48 184.75 252.7 185.12 252.71C185.49 252.72 185.79 252.5 185.8 252.26Z"
+              fill="#E5C699"
+            />
+            <path
+              id="Vector_262"
+              d="M188 255.28C188 254.78 187.45 254.35 186.75 254.33C186.05 254.31 185.46 254.7 185.44 255.2C185.42 255.7 185.99 256.12 186.69 256.14C187.39 256.16 187.94 255.78 188 255.28Z"
+              fill="#E5C699"
+            />
+            <path
+              id="Vector_263"
+              d="M143.22 239.28C143.445 239.136 143.618 238.924 143.714 238.675C143.811 238.426 143.825 238.152 143.756 237.894C143.687 237.636 143.538 237.407 143.33 237.239C143.122 237.071 142.866 236.974 142.599 236.961C142.333 236.949 142.069 237.021 141.846 237.168C141.623 237.315 141.452 237.529 141.359 237.779C141.265 238.03 141.254 238.303 141.326 238.56C141.398 238.817 141.55 239.045 141.76 239.21C141.948 239.407 142.204 239.524 142.476 239.537C142.748 239.55 143.014 239.458 143.22 239.28Z"
+              fill="#E5C699"
+            />
+            <path
+              id="Vector_264"
+              d="M145.7 236.94C145.78 236.69 145.57 236.39 145.23 236.27C144.89 236.15 144.53 236.27 144.44 236.49C144.35 236.71 144.57 237.03 144.91 237.16C145.045 237.228 145.201 237.243 145.347 237.203C145.493 237.162 145.619 237.068 145.7 236.94Z"
+              fill="#E5C699"
+            />
+            <path
+              id="Vector_265"
+              d="M149.66 226.44C149.75 226.19 149.54 225.9 149.2 225.77C149.065 225.702 148.909 225.687 148.762 225.727C148.616 225.768 148.491 225.862 148.41 225.99C148.32 226.24 148.53 226.54 148.88 226.66C149.23 226.78 149.58 226.68 149.66 226.44Z"
+              fill="#E5C699"
+            />
+            <path
+              id="Vector_266"
+              d="M133.71 230.65C133.79 230.41 133.58 230.11 133.24 229.98C133.105 229.912 132.949 229.897 132.803 229.937C132.657 229.978 132.531 230.072 132.45 230.2C132.37 230.45 132.58 230.75 132.92 230.87C133.26 230.99 133.62 230.9 133.71 230.65Z"
+              fill="#E5C699"
+            />
+            <path
+              id="Vector_267"
+              d="M109.87 211.5C109.96 211.25 109.75 210.95 109.4 210.83C109.05 210.71 108.71 210.83 108.62 211.05C108.53 211.27 108.74 211.59 109.09 211.72C109.224 211.788 109.378 211.803 109.523 211.762C109.667 211.721 109.791 211.628 109.87 211.5Z"
+              fill="#D8A96B"
+            />
+            <path
+              id="Vector_268"
+              d="M143 231.5C143.119 231.405 143.197 231.269 143.22 231.119C143.242 230.968 143.206 230.815 143.12 230.69C142.96 230.37 142.63 230.19 142.4 230.31C142.283 230.406 142.206 230.542 142.184 230.692C142.162 230.841 142.196 230.994 142.28 231.12C142.43 231.5 142.75 231.66 143 231.5Z"
+              fill="#E5C699"
+            />
+            <path
+              id="Vector_269"
+              d="M141.37 234.88C141.597 234.698 141.747 234.438 141.79 234.15C141.833 233.863 141.765 233.57 141.6 233.33C141.28 232.7 140.66 232.33 140.21 232.59C139.982 232.773 139.832 233.036 139.789 233.325C139.747 233.614 139.815 233.909 139.98 234.15C140.3 234.78 140.92 235.1 141.37 234.88Z"
+              fill="#E5C699"
+            />
+            <path
+              id="Vector_270"
+              d="M114.85 216.36C115.078 216.215 115.254 216.002 115.352 215.751C115.451 215.5 115.466 215.224 115.397 214.963C115.328 214.702 115.178 214.47 114.968 214.3C114.758 214.13 114.5 214.032 114.23 214.019C113.961 214.006 113.694 214.079 113.469 214.228C113.244 214.377 113.072 214.594 112.979 214.847C112.885 215.1 112.874 215.376 112.948 215.636C113.022 215.895 113.177 216.124 113.39 216.29C113.579 216.485 113.835 216.6 114.106 216.613C114.377 216.626 114.643 216.536 114.85 216.36Z"
+              fill="#E5C699"
+            />
+            <path
+              id="Vector_271"
+              d="M117.33 214.01C117.41 213.77 117.2 213.47 116.86 213.34C116.724 213.272 116.569 213.257 116.422 213.298C116.276 213.338 116.151 213.432 116.07 213.56C115.99 213.81 116.2 214.11 116.54 214.23C116.88 214.35 117.24 214.26 117.33 214.01Z"
+              fill="#E5C699"
+            />
+            <path
+              id="Vector_272"
+              d="M114.61 208.62C114.729 208.525 114.807 208.389 114.83 208.239C114.852 208.088 114.816 207.935 114.73 207.81C114.57 207.48 114.24 207.31 114.01 207.43C113.894 207.524 113.817 207.659 113.795 207.807C113.773 207.955 113.807 208.106 113.89 208.23C114.06 208.56 114.38 208.73 114.61 208.62Z"
+              fill="#E5C699"
+            />
+            <path
+              id="Vector_273"
+              d="M113 211.95C113.227 211.769 113.377 211.508 113.42 211.22C113.463 210.933 113.395 210.64 113.23 210.4C112.91 209.77 112.29 209.4 111.84 209.67C111.613 209.851 111.463 210.112 111.42 210.4C111.378 210.687 111.445 210.98 111.61 211.22C111.93 211.85 112.55 212.18 113 211.95Z"
+              fill="#E5C699"
+            />
+            <path
+              id="Vector_274"
+              d="M130.34 233.98C130.34 233.72 130.13 233.46 129.76 233.4C129.39 233.34 129.07 233.49 129.03 233.75C128.99 234.01 129.24 234.26 129.6 234.33C129.96 234.4 130.29 234.23 130.34 233.98Z"
+              fill="#E5C699"
+            />
+            <path
+              id="Vector_275"
+              d="M126.76 229.11C126.97 228.96 126.96 228.59 126.76 228.3C126.56 228.01 126.2 227.88 125.99 228.04C125.78 228.2 125.79 228.56 125.99 228.85C126.19 229.14 126.55 229.26 126.76 229.11Z"
+              fill="#E5C699"
+            />
+            <path
+              id="Vector_276"
+              d="M125.73 232.67C125.927 232.456 126.036 232.176 126.036 231.885C126.036 231.595 125.927 231.314 125.73 231.1C125.31 230.53 124.64 230.31 124.24 230.61C124.039 230.822 123.927 231.103 123.927 231.395C123.927 231.687 124.039 231.968 124.24 232.18C124.66 232.75 125.32 232.97 125.73 232.67Z"
+              fill="#E5C699"
+            />
+            <path
+              id="Vector_277"
+              d="M197.27 253.06C197.27 252.56 196.72 252.14 196.02 252.12C195.32 252.1 194.72 252.48 194.71 252.98C194.7 253.48 195.25 253.9 195.96 253.93C196.67 253.96 197.25 253.56 197.27 253.06Z"
+              fill="#E5C699"
+            />
+            <path
+              id="Vector_278"
+              d="M225.46 238.77C225.263 238.583 225.002 238.479 224.73 238.479C224.458 238.479 224.197 238.583 224 238.77C223.47 239.15 223.26 239.77 223.53 240.14C223.8 240.51 224.45 240.51 224.99 240.14C225.226 240.012 225.405 239.8 225.492 239.547C225.579 239.293 225.568 239.016 225.46 238.77Z"
+              fill="#E5C699"
+            />
+            <path
+              id="Vector_279"
+              d="M231.36 238.39C231.026 238.078 230.587 237.904 230.13 237.904C229.673 237.904 229.233 238.078 228.9 238.39C228.01 239.03 227.65 240.08 228.11 240.71C228.442 241.024 228.882 241.2 229.34 241.2C229.797 241.2 230.238 241.024 230.57 240.71C231.46 240.07 231.82 239.03 231.36 238.39Z"
+              fill="#D8A96B"
+            />
+            <path
+              id="Vector_280"
+              d="M221.61 242.05C221.275 241.74 220.836 241.568 220.38 241.568C219.924 241.568 219.485 241.74 219.15 242.05C218.25 242.7 217.9 243.74 218.36 244.38C218.693 244.689 219.131 244.861 219.585 244.861C220.039 244.861 220.477 244.689 220.81 244.38C221.71 243.73 222.07 242.69 221.61 242.05Z"
+              fill="#D8A96B"
+            />
+            <path
+              id="Vector_281"
+              d="M223.9 235.74C223.68 235.58 223.34 235.74 223.12 236C223.017 236.11 222.959 236.255 222.959 236.405C222.959 236.556 223.017 236.701 223.12 236.81C223.33 236.97 223.67 236.81 223.89 236.55C224.11 236.29 224.11 235.89 223.9 235.74Z"
+              fill="#E5C699"
+            />
+            <path
+              id="Vector_282"
+              d="M218 236.84C218 236.58 217.62 236.43 217.26 236.5C216.9 236.57 216.64 236.83 216.69 237.08C216.74 237.33 217.07 237.49 217.43 237.42C217.79 237.35 218 237.09 218 236.84Z"
+              fill="#E5C699"
+            />
+            <path
+              id="Vector_283"
+              d="M220.71 239.32C220.62 238.83 219.98 238.53 219.29 238.66C218.6 238.79 218.1 239.3 218.2 239.79C218.3 240.28 218.93 240.57 219.62 240.44C220.31 240.31 220.8 239.81 220.71 239.32Z"
+              fill="#E5C699"
+            />
+            <path
+              id="Vector_284"
+              d="M250 211.5C249.54 211.5 249.13 211.97 249.09 212.63C249.05 213.29 249.39 213.84 249.85 213.87C250.31 213.9 250.72 213.39 250.77 212.74C250.82 212.09 250.5 211.5 250 211.5Z"
+              fill="#E5C699"
+            />
+            <path
+              id="Vector_285"
+              d="M246.7 210.75C246.45 210.81 246.32 211.16 246.41 211.51C246.5 211.86 246.78 212.1 247.03 212.04C247.28 211.98 247.41 211.63 247.32 211.27C247.23 210.91 247 210.68 246.7 210.75Z"
+              fill="#E5C699"
+            />
+            <path
+              id="Vector_286"
+              d="M247.44 215.5C247 215.26 246.37 215.57 246.02 216.19C245.845 216.422 245.766 216.713 245.8 217.001C245.833 217.29 245.977 217.555 246.2 217.74C246.64 217.99 247.27 217.68 247.61 217.06C247.782 216.826 247.86 216.536 247.829 216.247C247.797 215.958 247.659 215.692 247.44 215.5Z"
+              fill="#E5C699"
+            />
+            <path
+              id="Vector_287"
+              d="M229.33 235.15C229.24 234.66 228.6 234.37 227.9 234.5C227.2 234.63 226.72 235.13 226.81 235.63C226.9 236.13 227.54 236.41 228.24 236.28C228.94 236.15 229.42 235.65 229.33 235.15Z"
+              fill="#E5C699"
+            />
+            <path
+              id="Vector_288"
+              d="M225.39 229.29C225.56 228.82 225.16 228.24 224.5 228C223.84 227.76 223.16 227.94 222.99 228.42C222.82 228.9 223.22 229.42 223.88 229.7C224.54 229.98 225.22 229.76 225.39 229.29Z"
+              fill="#E5C699"
+            />
+            <path
+              id="Vector_289"
+              d="M233.67 167.55C233.826 167.772 234.06 167.926 234.326 167.982C234.591 168.038 234.868 167.991 235.1 167.85C235.7 167.59 236.04 167.03 235.85 166.6C235.696 166.379 235.463 166.225 235.199 166.169C234.936 166.113 234.66 166.16 234.43 166.3C233.83 166.56 233.49 167.12 233.67 167.55Z"
+              fill="#D8A96B"
+            />
+            <path
+              id="Vector_290"
+              d="M234.66 172.07C234.83 172.26 235.19 172.22 235.47 171.97C235.595 171.886 235.684 171.758 235.717 171.611C235.751 171.464 235.727 171.31 235.65 171.18C235.529 171.091 235.378 171.05 235.227 171.067C235.077 171.083 234.939 171.156 234.84 171.27C234.56 171.5 234.48 171.87 234.66 172.07Z"
+              fill="#E5C699"
+            />
+            <path
+              id="Vector_291"
+              d="M242.86 165.25C243.01 165.25 243.13 165.08 243.13 164.87C243.13 164.66 243 164.49 242.85 164.49C242.7 164.49 242.59 164.66 242.59 164.87C242.59 165.08 242.71 165.25 242.86 165.25Z"
+              fill="#E5C699"
+            />
+            <path
+              id="Vector_292"
+              d="M238.42 165.04C238.42 165.54 238.97 165.96 239.67 166.04C240.37 166.12 240.97 165.67 240.98 165.17C240.99 164.67 240.44 164.25 239.73 164.17C239.02 164.09 238.44 164.5 238.42 165.04Z"
+              fill="#E5C699"
+            />
+            <path
+              id="Vector_293"
+              d="M218.66 180.65C218.858 180.835 219.119 180.938 219.39 180.938C219.661 180.938 219.922 180.835 220.12 180.65C220.65 180.27 220.86 179.65 220.59 179.28C220.376 179.13 220.121 179.051 219.859 179.053C219.598 179.055 219.344 179.138 219.132 179.29C218.92 179.443 218.761 179.658 218.676 179.905C218.592 180.152 218.586 180.419 218.66 180.67V180.65Z"
+              fill="#E5C699"
+            />
+            <path
+              id="Vector_294"
+              d="M177.74 188.5C177.35 189.18 177.8 190.18 178.74 190.74C179.099 191.024 179.553 191.159 180.009 191.118C180.464 191.077 180.887 190.863 181.19 190.52C181.58 189.85 181.13 188.85 180.19 188.29C179.831 188.008 179.378 187.873 178.923 187.912C178.469 187.951 178.045 188.161 177.74 188.5Z"
+              fill="#E5C699"
+            />
+            <path
+              id="Vector_295"
+              d="M226.94 173.82C227.066 174.255 227.327 174.638 227.685 174.914C228.043 175.191 228.48 175.346 228.932 175.357C229.385 175.368 229.829 175.235 230.201 174.977C230.572 174.719 230.852 174.35 231 173.922C231.147 173.494 231.155 173.031 231.021 172.598C230.887 172.166 230.619 171.788 230.256 171.518C229.893 171.248 229.453 171.101 229.001 171.098C228.549 171.094 228.107 171.235 227.74 171.5C226.84 172.14 226.48 173.19 226.94 173.82Z"
+              fill="#D8A96B"
+            />
+            <path
+              id="Vector_296"
+              d="M220.22 183.68C220.43 183.84 220.78 183.68 221 183.42C221.102 183.308 221.158 183.162 221.158 183.01C221.158 182.858 221.102 182.712 221 182.6C220.79 182.45 220.45 182.6 220.23 182.87C220.01 183.14 220 183.5 220.22 183.68Z"
+              fill="#E5C699"
+            />
+            <path
+              id="Vector_297"
+              d="M143.06 176.89C143.27 177.04 143.62 176.89 143.83 176.62C143.897 176.575 143.952 176.515 143.991 176.444C144.029 176.374 144.05 176.295 144.051 176.214C144.053 176.133 144.035 176.054 143.999 175.982C143.963 175.91 143.91 175.847 143.845 175.8C143.779 175.753 143.704 175.722 143.624 175.71C143.544 175.699 143.463 175.706 143.387 175.733C143.311 175.76 143.242 175.804 143.187 175.863C143.132 175.922 143.092 175.993 143.07 176.07C142.965 176.18 142.905 176.326 142.903 176.478C142.901 176.631 142.957 176.778 143.06 176.89Z"
+              fill="#E5C699"
+            />
+            <path
+              id="Vector_298"
+              d="M118.3 164.98C118.51 165.13 118.85 164.98 119.07 164.71C119.29 164.44 119.29 164.05 119.07 163.9C118.933 163.836 118.777 163.828 118.633 163.876C118.49 163.925 118.371 164.026 118.3 164.16C118.196 164.271 118.138 164.418 118.138 164.57C118.138 164.722 118.196 164.869 118.3 164.98Z"
+              fill="#E5C699"
+            />
+            <path
+              id="Vector_299"
+              d="M137.63 169.35C137.767 169.414 137.924 169.422 138.067 169.374C138.21 169.326 138.329 169.224 138.4 169.09C138.504 168.979 138.562 168.833 138.562 168.68C138.562 168.528 138.504 168.382 138.4 168.27C138.19 168.12 137.84 168.27 137.63 168.54C137.524 168.649 137.465 168.794 137.465 168.945C137.465 169.097 137.524 169.242 137.63 169.35Z"
+              fill="#E5C699"
+            />
+            <path
+              id="Vector_300"
+              d="M228.14 177.37C228.14 177.62 228.52 177.78 228.88 177.71C229.24 177.64 229.49 177.38 229.44 177.12C229.39 176.86 229.06 176.72 228.7 176.78C228.34 176.84 228.09 177.11 228.14 177.37Z"
+              fill="#E5C699"
+            />
+            <path
+              id="Vector_301"
+              d="M222.85 176.33C222.95 176.82 223.58 177.12 224.28 176.99C224.98 176.86 225.46 176.35 225.37 175.86C225.28 175.37 224.64 175.08 223.94 175.21C223.24 175.34 222.76 175.84 222.85 176.33Z"
+              fill="#E5C699"
+            />
+            <path
+              id="Vector_302"
+              d="M214.79 184.26C214.88 184.76 215.52 185.05 216.22 184.92C216.92 184.79 217.4 184.28 217.31 183.79C217.22 183.3 216.58 183.01 215.88 183.14C215.18 183.27 214.7 183.77 214.79 184.26Z"
+              fill="#E5C699"
+            />
+            <path
+              id="Vector_303"
+              d="M124.32 157.12C124.25 157.58 124.73 158.03 125.38 158.12C126.03 158.21 126.61 157.91 126.67 157.45C126.73 156.99 126.26 156.55 125.61 156.45C124.96 156.35 124.38 156.66 124.32 157.12Z"
+              fill="#E5C699"
+            />
+            <path
+              id="Vector_304"
+              d="M123.34 160.39C123.34 160.65 123.72 160.8 124.08 160.74C124.44 160.68 124.7 160.41 124.65 160.15C124.6 159.89 124.27 159.75 123.91 159.81C123.55 159.87 123.3 160.14 123.34 160.39Z"
+              fill="#E5C699"
+            />
+            <path
+              id="Vector_305"
+              d="M119 157.05C118.73 157.47 119 158.12 119.58 158.51C119.798 158.703 120.081 158.804 120.372 158.793C120.663 158.782 120.938 158.659 121.14 158.45C121.42 158.03 121.14 157.38 120.57 156.99C120.349 156.799 120.065 156.699 119.773 156.71C119.482 156.721 119.205 156.843 119 157.05Z"
+              fill="#E5C699"
+            />
+            <path
+              id="Vector_306"
+              d="M155.09 183.88C154.86 184.027 154.696 184.257 154.633 184.522C154.569 184.786 154.611 185.066 154.75 185.3C154.99 185.91 155.54 186.3 155.97 186.09C156.198 185.942 156.36 185.712 156.424 185.448C156.487 185.184 156.446 184.905 156.31 184.67C156.06 184.06 155.52 183.71 155.09 183.88Z"
+              fill="#E5C699"
+            />
+            <path
+              id="Vector_307"
+              d="M151.76 184.67C151.56 184.84 151.6 185.2 151.83 185.48C151.912 185.607 152.039 185.698 152.185 185.737C152.331 185.776 152.486 185.759 152.62 185.69C152.715 185.57 152.761 185.419 152.748 185.266C152.735 185.114 152.664 184.972 152.55 184.87C152.31 184.59 152 184.5 151.76 184.67Z"
+              fill="#E5C699"
+            />
+            <path
+              id="Vector_308"
+              d="M152.5 179.26C152 179.26 151.57 179.78 151.5 180.48C151.43 181.18 151.83 181.78 152.33 181.81C152.83 181.84 153.26 181.29 153.33 180.59C153.4 179.89 153 179.29 152.5 179.26Z"
+              fill="#E5C699"
+            />
+            <path
+              id="Vector_309"
+              d="M129.68 160.5C129.18 160.5 128.74 161.02 128.68 161.73C128.62 162.44 129.01 163.03 129.51 163.05C130.01 163.07 130.45 162.53 130.51 161.83C130.57 161.13 130.17 160.56 129.68 160.5Z"
+              fill="#E5C699"
+            />
+            <path
+              id="Vector_310"
+              d="M171.28 186.95C170.96 187.3 171.1 187.95 171.58 188.38C171.749 188.591 171.994 188.729 172.262 188.764C172.53 188.8 172.802 188.73 173.02 188.57C173.34 188.23 173.2 187.57 172.72 187.14C172.24 186.71 171.59 186.61 171.28 186.95Z"
+              fill="#E5C699"
+            />
+            <path
+              id="Vector_311"
+              d="M166.09 189.39C165.98 189.62 166.16 189.94 166.5 190.09C166.628 190.17 166.781 190.199 166.929 190.172C167.078 190.144 167.21 190.061 167.3 189.94C167.41 189.7 167.22 189.38 166.89 189.23C166.56 189.08 166.2 189.15 166.09 189.39Z"
+              fill="#E5C699"
+            />
+            <path
+              id="Vector_312"
+              d="M163 186.4C162.54 186.58 162.38 187.27 162.64 187.92C162.9 188.57 163.5 188.92 163.96 188.77C164.203 188.608 164.375 188.36 164.442 188.076C164.509 187.792 164.465 187.493 164.32 187.24C164.06 186.59 163.46 186.21 163 186.4Z"
+              fill="#E5C699"
+            />
+            <path
+              id="Vector_313"
+              d="M218.73 190.13C218.56 190.6 218.96 191.18 219.62 191.42C220.28 191.66 220.96 191.47 221.13 191C221.3 190.53 220.9 190 220.24 189.72C219.58 189.44 218.9 189.66 218.73 190.13Z"
+              fill="#E5C699"
+            />
+            <path
+              id="Vector_314"
+              d="M143.33 171.5C143.16 171.97 143.56 172.55 144.23 172.79C144.9 173.03 145.57 172.85 145.74 172.38C145.91 171.91 145.51 171.38 144.84 171.09C144.17 170.8 143.5 171.04 143.33 171.5Z"
+              fill="#E5C699"
+            />
+            <path
+              id="Vector_315"
+              d="M133.69 164.42C133.52 164.89 133.92 165.47 134.58 165.71C134.839 165.847 135.141 165.88 135.423 165.802C135.706 165.723 135.948 165.541 136.1 165.29C136.27 164.82 135.86 164.29 135.2 164.01C134.54 163.73 133.86 163.95 133.69 164.42Z"
+              fill="#E5C699"
+            />
+            <path
+              id="Vector_316"
+              d="M175.34 247.14C175.5 246.94 175.4 246.59 175.11 246.36C175.069 246.29 175.011 246.231 174.942 246.188C174.873 246.145 174.795 246.12 174.714 246.114C174.633 246.108 174.552 246.122 174.477 246.155C174.403 246.188 174.338 246.238 174.287 246.302C174.236 246.365 174.202 246.44 174.187 246.52C174.172 246.6 174.177 246.682 174.201 246.76C174.225 246.838 174.267 246.908 174.324 246.966C174.382 247.023 174.452 247.066 174.53 247.09C174.82 247.32 175.18 247.34 175.34 247.14Z"
+              fill="#E5C699"
+            />
+            <path
+              id="Vector_317"
+              d="M161.89 251.98C161.959 251.845 161.974 251.689 161.931 251.544C161.888 251.399 161.791 251.276 161.66 251.2C161.37 250.97 161.01 250.95 160.85 251.15C160.69 251.35 160.79 251.7 161.08 251.93C161.183 252.04 161.324 252.107 161.475 252.116C161.626 252.126 161.774 252.077 161.89 251.98Z"
+              fill="#E5C699"
+            />
+            <path
+              id="Vector_318"
+              d="M180.77 255.5C181.27 255.43 181.59 254.81 181.49 254.11C181.39 253.41 180.91 252.9 180.42 252.97C179.93 253.04 179.6 253.66 179.7 254.36C179.8 255.06 180.28 255.5 180.77 255.5Z"
+              fill="#E5C699"
+            />
+            <path
+              id="Vector_319"
+              d="M146.28 242.86C146.488 242.655 146.609 242.377 146.618 242.085C146.628 241.793 146.524 241.508 146.33 241.29C145.94 240.7 145.28 240.45 144.87 240.73C144.662 240.935 144.541 241.213 144.531 241.505C144.522 241.797 144.625 242.082 144.82 242.3C145.21 242.88 145.87 243.14 146.28 242.86Z"
+              fill="#E5C699"
+            />
+            <path
+              id="Vector_320"
+              d="M153.22 220.5C152.85 220.41 149.86 221.32 146.45 223.91C146.89 231.2 148.34 236.56 150.05 236.56C151.76 236.56 153.42 230.44 153.73 222.38C153.54 221.18 153.36 220.5 153.22 220.5Z"
+              fill="#F4EBCD"
+            />
+            <path
+              id="Vector_321"
+              d="M177.12 240.22C176.53 240.15 175.62 241.02 174.63 242.49C174.93 246.3 175.9 249.08 177.05 249.08C177.99 249.08 178.81 247.22 179.25 244.46C178.48 242.06 177.71 240.29 177.12 240.22Z"
+              fill="#F4EBCD"
+            />
+            <path
+              id="Vector_322"
+              d="M248.62 140.33C243.52 140.55 238.86 150.33 220.47 171.14C202.08 191.95 170.67 182.14 161.81 176.39C152.95 170.64 120.14 135.68 117 135.9C113.86 136.12 90.81 146.5 87.49 152.74C84.17 158.98 78.18 185.32 82.39 202.17C86.6 219.02 130.65 257.75 148.21 267.99C156.21 272.65 189.21 276.64 196.09 277.08C202.97 277.52 219.14 270.65 235.54 258.91C251.94 247.17 267.9 198.84 269 186.43C270.1 174.02 253.71 140.11 248.62 140.33ZM250 215.5C244.9 225.69 209.24 267.66 170.87 257.39C142.72 249.85 123.44 229.9 116.79 223.92C110.14 217.94 102.61 201.98 100.79 191.78C98.97 181.58 104.37 152.92 116.79 150.56C121.45 149.67 140.73 165.63 146.93 174.72C153.13 183.81 187.27 194.66 205.93 185.58C224.59 176.5 239.84 158.09 246.49 160.09C253.14 162.09 258.73 197.9 250 215.5Z"
+              fill="#A763CC"
+            />
+            <path
+              id="Vector_323"
+              d="M234.28 234.5L233.28 235.5L233.21 235.57L232.16 236.57L232.07 236.65L231 237.65L230.91 237.73L229.81 238.73L229.72 238.81L228.6 239.76L228.5 239.84L227.36 240.78L227.25 240.87L226.08 241.81L225.99 241.88L224.78 242.82L224.69 242.89L223.44 243.82L223.37 243.87C222.95 244.19 222.52 244.5 222.1 244.8L222.03 244.85L220.72 245.77H220.66L219.31 246.68L217.93 247.58L216.51 248.47C208.3 253.52 199.05 257.42 189.29 258.6C194.49 264.15 205.29 265.14 216.18 260.39C229.32 254.67 237.8 242.76 235.51 233.48L234.51 234.48L234.28 234.5Z"
+              fill="#CC80E0"
+            />
+            <path
+              id="Vector_324"
+              d="M170.82 257.36C164.47 255.634 158.312 253.269 152.44 250.3C153.22 255.63 158.27 260.84 165.44 263.02C173.04 265.34 180.55 263.5 183.93 258.95C179.507 259.033 175.095 258.498 170.82 257.36Z"
+              fill="#CC80E0"
+            />
+            <path
+              id="Vector_325"
+              d="M107.81 211.13C105.86 210.45 104.44 210.38 103.81 211.05C101.5 213.32 103.64 223.45 114.81 233.42C125.98 243.39 141.88 249.42 144.18 247.14C144.332 246.979 144.449 246.789 144.523 246.58C144.597 246.372 144.626 246.15 144.61 245.93C130.73 237.5 121.11 227.77 116.8 223.93C113.74 221.19 110.56 216.5 107.81 211.13Z"
+              fill="#CC80E0"
+            />
+            <path
+              id="Vector_326"
+              d="M108.69 155.6L109.86 155.11C111.544 152.823 113.98 151.201 116.74 150.53C117.192 150.468 117.652 150.502 118.09 150.63C121.144 148.595 123.898 146.141 126.27 143.34C121.51 138.88 117.96 135.83 117.04 135.9C113.94 136.12 90.8101 146.54 87.4901 152.74C86.5844 154.608 85.8545 156.557 85.3101 158.56C92.0001 159.84 100.34 158.95 108.69 155.6Z"
+              fill="#CC80E0"
+            />
+            <path
+              id="Vector_327"
+              d="M140.64 167.39C142.927 169.639 145.015 172.081 146.88 174.69C150.99 180.69 167.33 187.48 183.3 188.76C186.355 187.778 189.123 186.061 191.36 183.76C178.62 184 166.73 179.56 161.81 176.36C157.38 173.48 146.97 163.36 137.28 153.89C137.314 158.593 138.466 163.22 140.64 167.39Z"
+              fill="#CC80E0"
+            />
+            <path
+              id="Vector_328"
+              d="M221.13 175.96C232.63 167.47 241.69 158.64 246.4 160.06C246.711 160.158 247.002 160.31 247.26 160.51C250.316 156.399 252.6 151.767 254 146.84C254.09 146.53 254.16 146.22 254.24 145.92C251.95 142.45 249.91 140.28 248.61 140.33C243.51 140.55 238.85 150.33 220.46 171.14C219.045 172.733 217.496 174.202 215.83 175.53C217.026 175.776 218.24 175.923 219.46 175.97C220 175.99 220.58 175.98 221.13 175.96Z"
+              fill="#CC80E0"
+            />
+            <path
+              id="Vector_329"
+              d="M250 215.5C248.487 218.235 246.723 220.822 244.73 223.23C255.22 220.04 262.46 201.87 261.49 187.41C260.89 178.41 257.11 172.91 252.39 171.91C255.24 184.41 255.73 203.9 250 215.5Z"
+              fill="#CC80E0"
+            />
+            <path
+              id="Vector_330"
+              d="M247.51 156.29C241.58 154.57 229.45 167.07 213.97 176.46C211.53 178.19 205.48 182.82 202.74 183.55C175.48 190.87 150.2 178.08 143.74 169C137.28 159.92 120.74 141.47 115.94 142.35C103.13 144.71 95.9399 173.13 97.7199 183.35C98.5804 187.594 99.8912 191.734 101.63 195.7C101.26 194.33 100.96 193.01 100.74 191.77C98.9699 181.58 104.74 153.77 117.15 151.43C121.8 150.55 140.64 165.62 146.84 174.71C153.04 183.8 187.18 194.65 205.84 185.57C224.5 176.49 239.53 158.86 246.18 160.85C251.59 162.48 256.58 185.73 253.44 204.1C260.45 186.18 254 158.18 247.51 156.29Z"
+              fill="#72369C"
+            />
+            <path
+              id="Vector_331"
+              d="M237.2 250.5C220.2 262.24 203.49 269.11 196.38 268.67C189.27 268.23 155.1 264.24 146.84 259.58C130.91 250.58 93.9301 219.74 82.1001 200.91C82.1901 201.33 82.2801 201.76 82.3901 202.18C86.6001 219.02 130.65 257.76 148.21 268C156.21 272.66 189.21 276.65 196.09 277.09C202.97 277.53 219.14 270.66 235.54 258.92C248.1 249.92 260.41 219.46 265.97 200.12C259.47 218.91 248.43 242.71 237.2 250.5Z"
+              fill="#72369C"
+            />
+            <path
+              id="Vector_332"
+              d="M171 180.3C178.5 181.12 185.67 182.76 191.46 182.52C203.98 181.99 211.41 178.31 211.41 178.31C211.41 178.31 206.31 183.18 191.09 183.63C184.251 183.697 177.452 182.57 171 180.3Z"
+              fill="#72369C"
+            />
+            <path
+              id="Vector_333"
+              d="M112.56 150.67C113.56 141.72 112.56 137.6 112.56 137.6L115.75 148.87L112.56 150.67Z"
+              fill="#72369C"
+            />
+            <path
+              id="Vector_334"
+              d="M108.46 155.59C105.58 147.06 102.9 143.76 102.9 143.76L110.55 152.62L108.46 155.59Z"
+              fill="#72369C"
+            />
+            <path
+              id="Vector_335"
+              d="M105.78 160.21C99.4701 156.76 86.6001 155.63 86.6001 155.63L106.8 157.11L105.78 160.21Z"
+              fill="#72369C"
+            />
+          </g>
+        </motion.g>
+      </g>
+      <defs>
+        <radialGradient
+          id="paint0_radial_69_18457"
+          cx={0}
+          cy={0}
+          r={1}
+          gradientUnits="userSpaceOnUse"
+          gradientTransform="translate(174.6 133.017) scale(216.046 216.232)"
+        >
+          <stop stopColor="#72369C" />
+          <stop offset={0.32} stopColor="#6F3499" />
+          <stop offset={0.56} stopColor="#67308F" />
+          <stop offset={0.78} stopColor="#58287F" />
+          <stop offset={0.98} stopColor="#441C68" />
+          <stop offset={1} stopColor="#411B65" />
+        </radialGradient>
+        <radialGradient
+          id="paint1_radial_69_18457"
+          cx={0}
+          cy={0}
+          r={1}
+          gradientUnits="userSpaceOnUse"
+          gradientTransform="translate(35.2618 185.819) scale(21.2616 16.5368)"
+        >
+          <stop stopColor="#622323" />
+          <stop offset={1} stopColor="#471212" />
+        </radialGradient>
+        <radialGradient
+          id="paint2_radial_69_18457"
+          cx={0}
+          cy={0}
+          r={1}
+          gradientUnits="userSpaceOnUse"
+          gradientTransform="translate(313.778 187.856) scale(21.0456 16.3688)"
+        >
+          <stop stopColor="#622323" />
+          <stop offset={1} stopColor="#471212" />
+        </radialGradient>
+        <clipPath id="clip0_69_18457">
+          <rect
+            width={320}
+            height={300}
+            fill="white"
+            transform="translate(15 40)"
+          />
+        </clipPath>
+        <clipPath id="clip1_69_18457">
+          <rect
+            width={200}
+            height={145}
+            fill="white"
+            transform="translate(75 135.5)"
+          />
+        </clipPath>
+      </defs>
+    </svg>
+  );
+};
