+++ src/shared/ui/parallax-text/index.ts
