+++ src/shared/ui/parallax-text/parallax-text.styles.css
