+++ src/shared/ui/parallax-text/parallax-text.types.ts
@@ -1,0 +1,4 @@
+export interface ParallaxProps {
+  children: string;
+  baseVelocity?: number;
+}
