--- src/widgets/parallax-text/parallax-text.types.ts
@@ -1,4 +1,0 @@
-export type ParallaxProps = {
-  children: string;
-  baseVelocity: number;
-};
